% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday}
\usage{
birthday(n)
}
\arguments{
\item{n}{a quantiative vector}
}
\value{
The probability that two people share the same birthday
}
\description{
Birthday
}
\examples{
birthday(1:20)
}
