% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBoot.R
\name{boot2pval}
\alias{boot2pval}
\title{Bootstrap interval for a two sample test}
\usage{
boot2pval(x1, x2, conf.level = 0.95, iter = 3000, mudiff = 0, test = "two")
}
\arguments{
\item{x1}{The first sample vector}

\item{x2}{The second sample vector}

\item{conf.level}{The confidence interval}

\item{iter}{Number of iterations}

\item{mudiff}{The differences of the mean}

\item{test}{Test regions}
}
\value{
A histogram and the P-value
}
\description{
A function that takes in two sets of data and a number of other parameters to produce a boostrap interval
}
\examples{
x = rnorm(30, mean = 10, sd = 12)
y = rnorm(30, mean = 6, sd = 8)
boot2pval(x, y, 0.99, 10000, 4, "two")
}
