% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyChebishev.R, R/MyFunction1.R
\name{chebishev}
\alias{chebishev}
\title{Chebishev function}
\usage{
chebishev(k)

chebishev(k)
}
\arguments{
\item{k}{number of standard deviations.}
}
\value{
a number of points, n, representing that, for any distribution, there are this many or a greater number of points within k standard deviations

The minimum number of points within k standard deviations.
}
\description{
This takes in a number of standard deviations, k, and returns the minimum number of points within that many standard deviations for any distribution

More detailed description.
}
\examples{
k = 2; chebishev(k)
chebishev(2)

}
