% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{fahrenheit to celsius}
\usage{
fahrenheit_to_celsius(temp_F)
}
\arguments{
\item{tempF}{}
}
\value{
tempC
}
\description{
fahrenheit to celsius
}
\examples{
MATH4753::fahrenheit_to_celsiuc(32)
return 0
}
