% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchCSV.R
\name{FetchCSV}
\alias{FetchCSV}
\title{FetchCSV}
\usage{
FetchCSV(path)
}
\arguments{
\item{path}{string that represents directory path}
}
\value{
Returns a data table using data frame
}
\description{
FetchCSV
}
\examples{
   \dontrun{df = FetchCSV("/path/to/file.csv")}
}
