% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIREDAM.R
\docType{data}
\name{FIREDAM}
\alias{FIREDAM}
\title{Fire Insurance Company Fire Damage Data}
\format{
## `FIREDAM`
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{The distance of the fire from the fire station in miles}
  \item{DAMAGE}{Fire damage in thousands of dollars}
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
FIREDAM
}
\description{
A study of data used by fire insurance companies documenting
15 recent fires in a large suburb of a major city serviced by a particular
fire station
}
\keyword{datasets}
