% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{Get Mode}
\usage{
getmode(v)
}
\arguments{
\item{v}{V is a vector of values}
}
\value{
A number that is the mode of the vector.
}
\description{
Get Mode
}
\examples{
\dontrun{getmode(v)}
}
