% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab1.R
\name{lab1}
\alias{lab1}
\title{Mean and Standard Deviation from a vector
compute mean and standard deviation from a given vector}
\usage{
lab1(x)
}
\arguments{
\item{x}{a vector of interest}
}
\value{
a list of mean and standard deviation
}
\description{
Mean and Standard Deviation from a vector
compute mean and standard deviation from a given vector
}
\examples{
x = c(1,6,3,1,8); lab1(x)
}
