% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3.R
\name{lab3}
\alias{lab3}
\title{The Sum or Squares of SLR
Compute residual, model, and total sum of squares}
\usage{
lab3(x, y)
}
\arguments{
\item{x}{independent sample data}

\item{y}{random variable corresponding to x}
}
\value{
a list containing RSS, MSS, and TSS
}
\description{
The Sum or Squares of SLR
Compute residual, model, and total sum of squares
}
\examples{
x = df1$COLUMN1; y = df1$COLUMN2; lab3(x,y)
}
