% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab7.R
\name{lab7chisim2}
\alias{lab7chisim2}
\title{Two Population Chi Sq Statistics}
\usage{
lab7chisim2(
  n1 = 10,
  n2 = 14,
  sigma1 = 3,
  sigma2 = 3,
  mean1 = 5,
  mean2 = 10,
  iter = 1000,
  xleg = 12,
  ymax = 0.07,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Two Population Chi Sq Statistics
}
\examples{
lab7chisim2(n1=10,n2=14,sigma1=3,sigma2=3,mean1=5,mean2=10,iter=1000, ...)
}
