% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxlikg2.R
\name{maxlikg2}
\alias{maxlikg2}
\title{Maximum Likelihood for Two Estimators}
\usage{
maxlikg2(theta1, theta2, lfun = "logbinpois", ...)
}
\arguments{
\item{theta1}{First estimator values}

\item{theta2}{Second Estimator Values}

\item{lfun}{log likelihood function}

\item{...}{Additional parameters for countor map}
}
\value{
Contour map and list of estimates for theta1 and theta2
}
\description{
Maximum Likelihood for Two Estimators
}
\examples{
\dontrun{maxlikg2(x, y, lfun = logbinpois)}
}
