% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{sample size}

\item{p}{probability}
}
\value{
A barplot with extra colors and stuff
}
\description{
Binomial Function
}
\examples{
mybin(iter=20, n=10, p=0.7)
}
