% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybinroxygen.R
\name{mybinroxygen}
\alias{mybinroxygen}
\title{my bin roxygen}
\usage{
mybinroxygen(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{iterations}

\item{n}{sample size}

\item{p}{probability}
}
\value{
binomial experiment with plot and table
}
\description{
my bin roxygen
}
\examples{
\dontrun{mybin(iter=1000,n=18, p=0.3)}
}
