% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybiplot.R
\name{mybiplot}
\alias{mybiplot}
\title{A Bi-Plot}
\usage{
mybiplot(x, y, col)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
plots
}
\description{
A simple plotting function
}
\details{
This plot will take in two vectors and plot them
}
\examples{
x=1:30, y= x^2, mybiplot(x,y)
}
