% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{myboot}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{distribution}

\item{fun}{function}

\item{alpha}{1 - confidence interval in question}

\item{...}{}
}
\value{
A histogram of the input distribution displaying functions (mean, sd, IQR, etc) of the distribution and a confidence interval.
}
\description{
myboot
}
\examples{
\dontrun{myboot2(x = sam, fun = "var", alpha = .2)}
}
