% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap estimation}
\usage{
myboot2(
  iter = 10000,
  x,
  fun = "mean",
  alpha = 0.05,
  cx = 1.5,
  show = TRUE,
  ...
)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sampled data from population}

\item{fun}{statistic to calculate over x}

\item{alpha}{bootstrap interval delimiter, gives (1-alpha)*100 pct interval}

\item{cx}{character expansion factor for plotted intervals}

\item{show}{show the plot}

\item{...}{extra parameters to pass to graphics objects}
}
\value{
A histogram of sampled statistics, the estimated population statistic, and the bootstrap intervals. Also returns list containing confidence intervals, function used, sample set, and calculated statistics
}
\description{
Bootstrap estimation
}
\examples{
myboot2(x=rnorm(20,mean=10,sd=5),xlab="mean")
myboot2(x=rnorm(20,mean=10,sd=5),xlab="std deviation",fun="sd",col="lightblue")
}
