% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mychisim.R
\name{mychisim}
\alias{mychisim}
\title{Title}
\usage{
mychisim(
  lablex,
  labley,
  n1 = 10,
  sigma1 = 3,
  mean1 = 5,
  iter = 1000,
  ymax = 0.1,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
