% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab7func.R
\name{mychisim2}
\alias{mychisim2}
\title{mychisim2 RMD version from Lab 7}
\usage{
mychisim2(
  n1 = 10,
  n2 = 14,
  sigma1 = 3,
  sigma2 = 3,
  mean1 = 5,
  mean2 = 10,
  iter = 1000,
  ymax = 0.07,
  x = 40,
  y = 0.04,
  ...
)
}
\arguments{
\item{n1}{sample size for population 1}

\item{n2}{sample size for population 2}

\item{sigma1}{sd population 1}

\item{sigma2}{sd population 2}

\item{mean1}{mean for population 1}

\item{mean2}{mean for population 2}

\item{iter}{iterations}

\item{ymax}{Y max on the graphs}

\item{x}{vector x}

\item{y}{vector y}

\item{...}{...}
}
\value{
chi^2 graph for 2 populations
}
\description{
mychisim2 RMD version from Lab 7
}
\examples{
\dontrun{mychisim2(iter = 10000)}
}
