% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{My confidence interval generator}
\usage{
myci(y)
}
\arguments{
\item{y}{a vector of data to find a confidence interval on}
}
\value{
ci, a list containing the lower and upper boundaries of the interval

conflevel, the confidence level of the interval
}
\description{
Finds a 95 percent confidence interval for the mean
}
\details{
Given a set of data, computes and returns a 95 percent confidence interval for the mean of that set
}
\examples{
set.seed(12); sam1 = rnorm(20, mean = 5, sd = 3); myci(sam1)
}
