% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{sample size n}

\item{iter}{number of iterations}

\item{b}{}
}
\value{
Prints a histogram of the distribution of the sum of uniforms
}
\description{
myclt
}
\examples{
\dontrun{myclt(n = 10, iter = 10000)}
}
