% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltbin.R
\name{mycltb}
\alias{mycltb}
\title{mycltbin}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{number of times you want to do the trial}

\item{iter}{your iterations}

\item{p}{probability}

\item{...}{}
}
\value{
a graph
}
\description{
mycltbin
}
\examples{
mycltbin(10,1000,0.6) ;returns a graph
}
