% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltmv.R
\name{mycltmv}
\alias{mycltmv}
\title{myCLTmv}
\usage{
mycltmv(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{Size of each sample}

\item{iter}{The amount of samples to have}

\item{a}{The lower bound of the sampled uniform distribution}

\item{b}{The upper bound of the sampled uniform distribution}
}
\value{
A vector containing the sample means represented by the histogram
}
\description{
myCLTmv
}
\examples{
mycltmv(10, 10000, 0, 5)
}
