% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Title}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{sample size}

\item{iter}{iterations}

\item{a}{lower}

\item{b}{upper}
}
\value{
A barplot and curve of uniform distribution of sample size n, multiplied by iterations with parameters a and b
}
\description{
Title
}
\examples{
mycltu(n=20,iter=10000,a=0,b=15)

}
