% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{myddt}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{data file}

\item{SPECIES}{Species of fish}
}
\value{
Plots length vs. weight of species of fish. This function includes my name on top, is colored according to the river variable, has a quadratic curve placed over it, prints the subset data into a csv file in the working directory, and returns a list of the DDT data from before subsetting, after subsetting, and relative frequency.
}
\description{
myddt
}
\examples{
\dontrun{myddt(df = ddt, SPECIES = "CCATFISH")}
}
