% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{My Function of sqaures and cubes}
\usage{
myf(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a list of vectors
}
\description{
Produce list of squares and cubes
}
\details{
This isa part of the introduction to package package making
}
\examples{
x = 1:4; myf(x)
}
