% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFirstFunction.R
\name{myFirstFunction}
\alias{myFirstFunction}
\title{myFirstFunction}
\usage{
myFirstFunction(x)
}
\arguments{
\item{x}{A number}
}
\value{
A list of x and y, y is x^2
}
\description{
myFirstFunction
}
\examples{
myFirstFunction(1:20000)

}
