% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunddt.R
\name{myfunddt}
\alias{myfunddt}
\title{myfunddt}
\usage{
myfunddt(df, Species)
}
\arguments{
\item{df}{Data frame that is read into the function.}

\item{Species}{Specify the species of fish you want to plot data for. Options are LMBASS, CCATFISH, SMBUFFALO}
}
\value{
A graph for the ddt data frame. Plots Length vs Weight and adds a quadratic model.
}
\description{
myfunddt
}
\examples{
  \dontrun{myfunddt(df = ddt, Species  = "CCATFISH") }
}
