% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhead.R
\name{myhead}
\alias{myhead}
\title{Head of Data}
\usage{
myhead(obj)
}
\arguments{
\item{obj}{}
}
\value{
First 6 lines of code
}
\description{
This will return only the first 6 lines of data
}
\details{
It takes a given csv file and will read it then return only 6 lines of data.
}
\examples{
 obj = myread("Name.csv"); myhead(obj)
}
