% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{myhyper}
\alias{myhyper}
\title{hypergeometric simulation}
\usage{
myhyper(iter = 1000, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{Number of iterations}

\item{N}{N}

\item{r}{gamma}

\item{n}{n}
}
\value{
hyper graph
}
\description{
hypergeometric simulation
}
\examples{
\dontrun{myhyper(iter = 1000, N = 20, r = 12, n = 5)}
}
