% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{My Max Likelihood}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{likelihood function}

\item{x}{data}

\item{param}{vector of possible values for the parameter(s) of the likelihood function lfun that you want to estimate}

\item{...}{allows users to customize the plot}
}
\value{
returns a list with: i - The index of the parameter value where the maximum likelihood occurs. parami - The specific parameter value at index i where the maximum likelihood is found. yi - The maximum likelihood value from y[i]. slope - A vector of slopes around the maximum likelihood point
}
\description{
My Max Likelihood
}
