% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{Maximum likelihood of log binomial}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{log functions}

\item{theta}{theta}
}
\value{
Maximum likelihood of binomial experiments
}
\description{
Maximum likelihood of log binomial
}
\examples{
mymaxlikg(theta = seq(0,1,length=10000, lfun=logbin2))

}
