% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlbeta.R
\name{mymlbeta}
\alias{mymlbeta}
\title{Maximum Likelihood Estimation for Beta Distribution}
\usage{
mymlbeta(x, a_vals, b_vals, ...)
}
\arguments{
\item{x}{A numeric vector of values assumed to come from a Beta distribution.}

\item{a_vals}{A vector of possible values for shape1 (alpha).}

\item{b_vals}{A vector of possible values for shape2 (beta).}

\item{...}{Additional graphical parameters passed to \code{persp()}.}
}
\value{
A list containing:
\describe{
  \item{shape1_hat}{The estimated MLE for shape1 (alpha).}
  \item{shape2_hat}{The estimated MLE for shape2 (beta).}
  \item{loglik}{The value of the log-likelihood at the maximum.}
}
}
\description{
This function estimates the shape parameters of a beta distribution via maximum likelihood,
using a grid-based approach and returns the MLEs with a 3D log-likelihood plot.
}
\examples{
x <- rbeta(30, 3, 4)
a_vals <- seq(1, 6, length = 50)
b_vals <- seq(1, 6, length = 50)
result <- mymlbeta(x, a_vals, b_vals)
}
