% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlNorm}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{a vector, your trials}

\item{mu}{an integer, your mean}

\item{sig}{an integer, your sd}

\item{...}{}
}
\value{
a graph which will show max likelyhood estimates of population mean and sd
}
\description{
mymlNorm
}
\examples{
mymlnorm=function(x = data,mu = 4,sig = 2,...)
}
