% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbin.R
\name{mynbin}
\alias{mynbin}
\title{mynbin - a function to calculate a binomial distribution}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{Number of trials until r successes are observed}

\item{r}{Number of successes to be observed}

\item{p}{Probability of success}
}
\value{
A negative binomial probability distribution
}
\description{
mynbin - a function to calculate a binomial distribution
}
\examples{
\dontrun{mynbin(10,3,0.4)}
}
