% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbinFunction.R
\name{mynbino}
\alias{mynbino}
\title{My Negative Binomial Distribution Function}
\usage{
mynbino(y, r, p)
}
\arguments{
\item{y}{a vector ofnumber of trials until the kth success}

\item{r}{a vector which shows the size of the sample}

\item{p}{a vector in which its value shows the probability}
}
\value{
a negative binomial distribution vector in the form of a vector object
}
\description{
Finds negative binomial distribution based on the parameters size, probabilities and the value of x. So, in other words, vecators are used to do the calculation
}
\details{
This is the second function in this package for the MATH4753 class. It takes the values of the vectors and plugs them into the equation. The answer shows the negative binomial distribution, which is a discrete probability distribution in relation to the number of successes in a sequence of Bernoulli trials that are independent and identically distributed, and that is before a specified number of failures occurs.
}
\examples{
y=12, r=5, p=0.3;mynbino(y,r,p)
}
