% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbinom.R
\name{mynbinom}
\alias{mynbinom}
\title{Alternative for dnbinom()}
\usage{
mynbinom(y, r, p)
}
\arguments{
\item{y}{Numeric value of y in the negative binomial distribution}

\item{r}{Numeric value of r in the negative binomial distribution}

\item{p}{Numeric value of p in the negative binomial distribution}
}
\description{
Doesn't require subtracting r from y for x value, created for lab 5
}
\examples{
dnbinom(10, size = 3, prob = 0.4)

}
