% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab6.R
\name{myncurvelab6}
\alias{myncurvelab6}
\title{Continuous, Normal Distribution Area}
\usage{
myncurvelab6(mu, sigma, x)
}
\arguments{
\item{mu}{mean}

\item{sigma}{the standard deviation}

\item{x}{vector of interest}
}
\value{
graphic showing the area of prescribed configuration of a normal distribution
}
\description{
Continuous, Normal Distribution Area
}
\examples{
myncurvelab6(10,1,7)
}
