% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynorm.R
\name{mynorm}
\alias{mynorm}
\title{course 4753}
\usage{
mynorm(x, mu, sig, ...)
}
\arguments{
\item{x}{make a contour plot based on the value of mu and variance}
}
\value{
a contour plot which y = variance and x = mean
}
\description{
make a contour plot based on the value of mu and variance.
}
\examples{
mynorm(x = c(10,12,13,15,12,11,10),
mu = seq(10,15, length = 1000),
sig = seq(0.1,4,length = 1000),
lwd = 2, labcex = 1)


}
