% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNRML.R
\name{myNRML}
\alias{myNRML}
\title{myNRML}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{an integer, your starting point}

\item{delta}{an integer, your derivative/how close you want it to be}

\item{llik, }{the function that you want to take ln of}

\item{xrange}{a vector, the range of your values}

\item{parameter}{, what you want to do the derivative respect to}
}
\value{
a graph that will show the maximum likely hood by doing the derivative with respect to a variable
}
\description{
myNRML
}
\examples{
myNRML(x0=1,delta=0.000001,llik=function(x) log(dpois(4,x)*dpois(6,x)*dpois(7,x)*dpois(6,x)*dpois(5,x)),xrange=c(0,30),parameter="lambda" )
}
