% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{This function calculates the height of the tree given BHDiameter}
\usage{
myplot(x)
}
\arguments{
\item{x}{a quantitative vector with BHDiameter}
}
\value{
a quantitative vector of height
}
\description{
This function calculates the height of the tree given BHDiameter
}
\examples{
myplot(1:10)

}
