% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{My bootstrap p-value}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{the calculated t-statistic}

\item{xmax}{the maximum x value to include on the plot}

\item{n}{the size of the sample used to calculate the t-statistic}

\item{alpha}{used to define the width of the confidence interval as 1-alpha}
}
\value{
q, quantile of the rejection region

p, the p-value
}
\description{
Calculates p-value, rejection regions, and plots the t-distribution for a given t-value.
}
\details{
Given a calculated t-value and the size of the sample used to calculate it, returns a plot of the t-distribution, the rejection quantile, and the p-value.
}
\examples{
t0 = 2;mypvalue(t0=t0)
}
