% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myquad.R
\name{myquad}
\alias{myquad}
\title{Quadratic function}
\usage{
myquad(x, lm_model)
}
\arguments{
\item{x}{numeric vector}

\item{lm_model}{linear regression model object}
}
\value{
A quadratic equation
}
\description{
Quadratic function
}
\examples{
\dontrun{myquad(x = 1:10, df.lm)}


}
