% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{My csv read function}
\usage{
myread(csv, dird)
}
\arguments{
\item{csv}{a string with the filename of the csv to read}

\item{dird}{a string with the filepath up to the file where the csv is}
}
\value{
a data frame of values
}
\description{
Reads a csv file into the workspace.
}
\details{
Given a path string and csv name, reads the csv into a data frame in the workspace.
}
\examples{

dird = "[directory path]"; csv = "[filename]"
myread(csv, dird)
}
