% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions Package.R
\name{mysd}
\alias{mysd}
\title{My Standard Deviation Function}
\usage{
mysd(x)
}
\arguments{
\item{x}{a vector in which its value is used to calculate the standard deviation}
}
\value{
a standard deviation vector on the form of a vector object
}
\description{
Creates standard deviation from vectors
}
\details{
This is the first function in this package for the MATH4753 class
}
\examples{
x=1:30;mysd(x)
}
