% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySimulation.R
\name{mySimulation}
\alias{mySimulation}
\title{Binomial simulation plotting function}
\usage{
mySimulation(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{number of time for iteration}

\item{n}{number of size}

\item{p}{probability}
}
\value{
A table along with a barplot of the simulation
}
\description{
A function that takes in the n and p value, enter the number of iter of desire, and it will simulate a binomial distribution and plots
}
\examples{
iter = 10,000, n = 5, p = 0.5 mybin(iter, n, p)
}
