% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPackage.R
\name{mysnd}
\alias{mysnd}
\title{My Standard Normal Distribution Functiom}
\usage{
mysnd(x)
}
\arguments{
\item{x}{a vector in which its value is used to calculate the standard normal distribution}
}
\value{
a standard normal distribution vector in the form of a vector object
}
\description{
Finds standard normal distribution from vectors
}
\details{
This is the first function in this package for the MATH4753 class. It takes the values of the vectors and plugs them into the equation. The answer shows the standard normal distribution, or in other words, how much are the data far away from the mean value.
}
\examples{
x=1:30;mysnd(x)
}
