% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysq.R
\name{mysq}
\alias{mysq}
\title{Mysquared Function}
\usage{
mysq(x)
}
\arguments{
\item{x}{number to be squared (as a vector)}
}
\value{
the square of x (as a vector)
}
\description{
Mysquared Function
}
\examples{
mysq(4)
mysq(1:10)
}
