% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTail.R
\name{myTail}
\alias{myTail}
\title{Tail of data}
\usage{
myTail(obj)
}
\arguments{
\item{obj}{}
}
\value{
last 6 lines of data
}
\description{
This will give you the last 6 lines of data from a csv file
}
\examples{
obj = myread("Name.csv"); myTail(obj)
}
