% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab7func.R
\name{myTsim}
\alias{myTsim}
\title{myTsim RMD version from Lab 7}
\usage{
myTsim(
  n1 = 10,
  sigma1 = 3,
  mean1 = 5,
  iter = 1000,
  ymax = 0.1,
  x = 2,
  y = 0.3,
  ...
)
}
\arguments{
\item{n1}{sample size}

\item{sigma1}{standard deviation}

\item{mean1}{mean}

\item{iter}{iterations}

\item{ymax}{Y max on the graph}

\item{x}{vector x}

\item{y}{vector y}

\item{...}{...}
}
\value{
T plot
}
\description{
myTsim RMD version from Lab 7
}
\examples{
\dontrun{myTsim(iter = 10000, ymax = 0.45)}
}
