% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab7func.R
\name{myTsim2}
\alias{myTsim2}
\title{myTsim2 RMD version from Lab 7}
\usage{
myTsim2(
  n1 = 10,
  n2 = 14,
  sigma1 = 3,
  sigma2 = 3,
  mean1 = 5,
  mean2 = 10,
  iter = 1000,
  ymax = 0.5,
  x = 2,
  y = 0.4,
  ...
)
}
\arguments{
\item{n1}{sample size for population 1}

\item{n2}{sample size for population 2}

\item{sigma1}{standard deviation for population 1}

\item{sigma2}{standard deviation for population 2}

\item{mean1}{mean for population 1}

\item{mean2}{mean for population 2}

\item{iter}{iterations}

\item{ymax}{Y max on the graphs}

\item{x}{vector x}

\item{y}{vector y}

\item{...}{...}
}
\value{
T graph for 2 populations
}
\description{
myTsim2 RMD version from Lab 7
}
\examples{
\dontrun{myTsim2(iter = 10000)}
}
