% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myZfunction.R
\name{myZfunction}
\alias{myZfunction}
\title{My Function producing z values}
\usage{
myZfunction(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a list of vectors
}
\description{
Produce vectors of z values
}
\examples{
x = 1:7; myZfunction(x)
}
