% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfunction.R
\name{newfunction}
\alias{newfunction}
\title{course 4753}
\usage{
newfunction(n, iter)
}
\arguments{
\item{x}{making a plot with the value of n.}
}
\value{
a histogram of the data provide.
}
\description{
making a plot with the value of n.
}
\examples{
newfunction(n=10,iter=1000)


}
