% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Number of Tickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats}

\item{gamma}{probability of plane being overbooked}

\item{p}{probability of someone showing up}
}
\value{
A list containing nd, nc, p, and gamma. Creates two graphs displaying nd - using discrete distribution, and nc - using normal approximation
}
\description{
Number of Tickets
}
\examples{
ntickets(400, 0.02, 0.95)
}
