% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.r
\name{pareto}
\alias{pareto}
\title{Pareto chart}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{numeric vector}

\item{mn}{Name for Pareto Chart}
}
\value{
numeric vector with a Pareto chart
}
\description{
Pareto chart
}
\examples{
data <- c(10, 15, 7, 25, 8)
pareto(data, mn = "Pareto Chart Example")

}
