% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbinomf.R
\name{pbinomf}
\alias{pbinomf}
\title{probability for binomial distribution function}
\usage{
pbinomf(q, size, prob)
}
\arguments{
\item{q}{quantile}

\item{size}{number of trials}

\item{prob}{probability of success}
}
\value{
a pbinom value
}
\description{
probability for binomial distribution function
}
\examples{
\dontrun{pbinomf(q = 8, size = 15, prob = 0.4)}
}
