% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInData.R
\name{readInData}
\alias{readInData}
\title{Read in csv}
\usage{
readInData(x)
}
\arguments{
\item{x}{Name of csv file}
}
\value{
the read csv
}
\description{
Read in csv
}
\examples{
\dontrun{readInData("SPRUCE.csv")}
}
