% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{ScatterHist}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{A quantitative vector}

\item{y}{A quantitative vector}

\item{xlab}{The label for x}

\item{ylab}{The label for y}
}
\value{
A scatter plot of x vs y with marginal histograms
}
\description{
ScatterHist
}
\examples{
scatterhist(x = 1:10, y = 11:20)
}
