% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Title  function creates a scatter plot given x vector and y vector}
\usage{
scatterPlot(x, y)
}
\arguments{
\item{x}{a vector of quantitative data}

\item{y}{a vector of quantitative data}
}
\value{
None
}
\description{
Title  function creates a scatter plot given x vector and y vector
}
\examples{
scatterPlot(1:10, 1:10)
}
