% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttestcalc.R
\name{t.test.calc}
\alias{t.test.calc}
\title{Title}
\usage{
\method{t}{test.calc}(x, alpha, mu)
}
\arguments{
\item{x}{a vector containing a sample}

\item{alpha}{the probability of the type I error}

\item{mu}{the population mean}
}
\value{
returns R A R being the interval with which the t-value is expected to lie between.
}
\description{
Title
}
\examples{
x = rnorm(n = 30, mean = 5, sd =10);t.test.calc(x, alpha = .05, mu = 5)
}
