% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttestcalc.R
\name{ttestcalc}
\alias{ttestcalc}
\title{t.test.calc()}
\usage{
ttestcalc(x, alpha, mu)
}
\arguments{
\item{x}{a vector containing a sample}

\item{alpha}{the probability of the type I error}

\item{mu}{the population mean}
}
\value{
returns R A R being the interval with which the t-value is expected to lie between.
}
\description{
RAR statistic that deterines whether the t-value of the sample passed in and returns the t-value and the t-value intervals
}
\examples{
x = rnorm(n = 30, mean = 5, sd =10);t.test.calc(x, alpha = .05, mu = 5)
}
