% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot2pval.R
\name{boot2pval}
\alias{boot2pval}
\title{boot2pval}
\usage{
boot2pval(x1, x2, conf.level = 0.95, iter = 3000, mudiff = 0, test = "two")
}
\arguments{
\item{x1}{Sample 1}

\item{x2}{Sample 2}

\item{conf.level}{The level of the confidence interval}

\item{iter}{The number of iterations for interval generation}

\item{mudiff}{The null value for the difference of sample means}

\item{test}{A value that moderates what is generated}
}
\value{
The p-value and a graph depicting the confidence interval
}
\description{
boot2pval
}
\examples{
set.seed(10);x=rnorm(20,mean=4,sd=2);  y=rnorm(20,mean=4,sd=2)
boot2pval(x1=x,x2=y)
}
