% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot2pval.R
\name{boot2pval}
\alias{boot2pval}
\title{A function that takes in two sets of data and a number of other parameters to produce a boostrap interval}
\usage{
boot2pval(x1, x2, conf.level = 0.95, iter = 3000, mudiff = 0, test = "two")
}
\arguments{
\item{test}{}
}
\value{
#' A histogram and the P-value
}
\description{
A function that takes in two sets of data and a number of other parameters to produce a boostrap interval
}
\examples{
x = rnorm(30, mean = 10, sd = 12)
y = rnorm(30, mean = 6, sd = 8)
boot2pval(x, y, 0.99, 10000, 4, "two")
}
