% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab8.R
\name{clt_pois}
\alias{clt_pois}
\title{Title Central Limit Theorem on Poisson Distribution
Produce plots pertaining the distribution of the sample mean from a population distributed as
Poisson distribution}
\usage{
clt_pois(n, iter, lambda = 10, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title Central Limit Theorem on Poisson Distribution
Produce plots pertaining the distribution of the sample mean from a population distributed as
Poisson distribution
}
\examples{
clt_pois(n = 50, iter = 100000, ...)
}
