% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Dataset}
\format{
A data frame with columns representing DDT measurements.
}
\usage{
data(ddt)
}
\description{
Example dataset containing DDT data.
}
\examples{
data(ddt)
head(ddt)
}
\keyword{datasets}
