% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Concentration Data}
\format{
A data frame with variables:
\describe{
  \item{DDT}{numeric vector of DDT concentration measurements}
}
}
\source{
Original CSV file: DDT.csv in data-raw/
}
\usage{
ddt
}
\description{
A dataset containing DDT concentration values used for Lab 9 bootstrap examples.
}
\keyword{datasets}
