% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Fish data set}
\format{
## `who`
A data frame with 144 rows and 6 columns:
\describe{
  \item{RIVER}{Where the fish was found}
  \item{MILE}{The mile of river where it was found}
  \item{SPECIES}{The species of fish which was caught}
  \item{LENGTH, WEIGHT}{The length and weight of the caught fish}
  \item{DDT}{DDT found in the fish}
  ...
}
}
\usage{
ddt
}
\description{
A report on fish and their given levels of DDT found inside of them
}
\keyword{datasets}
