% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Data}
\format{
A data frame with X rows and Y variables:
\describe{
  \item{DDT}{DDT concentration (numeric)}
  \item{...}{Other variables if present}
}
}
\source{
Source info or note
}
\usage{
ddt
}
\description{
Sample data on DDT concentrations.
}
\keyword{datasets}
