% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set for MPG}
\format{
A data set of 100 rows and 1 variable:
\describe{
  \item{MPG}{miles per gallon}
}
}
\source{
text book
}
\usage{
ddt
}
\description{
This data set contains 100 lines of measurements related to MPG (Miles Per Gallon).
}
\keyword{datasets}
