% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIREDAM.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM data sets for fires that occurred}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{distance, in units of miles}
  \item{DAMAGE}{damage, in dollars}
  ...
}
}
\source{
text Book
}
\usage{
fire
}
\description{
A dataset containing the data for 15 fires which document the damages sustained by a fire at a certain distance.
}
\keyword{datasets}
