% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM.csv Data}
\format{
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Distance to the fire}
  \item{DAMAGE}{Damage count for a fire}
}
}
\source{
\url{http://www.fire.com/firedamdata}
}
\usage{
fire
}
\description{
Contains information about the distance and damage of fires.
}
\keyword{datasets}
