% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire data set.}
\format{A data fram with 15 rows and 2 variables
\describe{
   \item{DISTANCE}{distance from fire station in miles}
   \item{DAMAGE}{x 1000 dollars in damage}
}}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A data set collected by
}
\keyword{datasets}
