% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire data}
\format{
## `fire`
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{Distance}
  \item{DAMAGE}{Damage}
  ...
}
}
\source{
<test>
}
\usage{
fire
}
\description{
A subset of data from the some fire damage report
}
\keyword{datasets}
