% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage}
\format{}
\source{
\url{https:]]www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A data set collected concerning home damage from fire
}
\examples{
\dontrun{
 firedam
}
}
\keyword{datasets}
