% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Damage and distance scales of fires.}
\format{A data frame with 15 rows and 2 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\usage{
fire
}
\description{
A dataset containing the distance a fire has spread and the ammount of damage it has done.
}
\keyword{datasets}
