% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{Distance from a fire station}
  \item{DAMAGE}{Cost of damage in thousands of dollars}
  ...
}
}
\source{
Statistics for Engineering and the Sciences
}
\usage{
fire
}
\description{
Data that relates distance from a fire station to the cost of the damage...
}
\keyword{datasets}
