% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fish caught in the tennessee river and its tributaries}
\format{
A data frame with 144 rows and 6 variables:
\describe{
  \item{RIVER}{Three letter abbreviation of the River}
  \item{MILE}{Mile count for where the fish was caught}
  ...
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A dataset with 6 variables two of which are quantitative.
Interest is in the amount of DDT in their flesh
}
\keyword{datasets}
