% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
A data frame with N rows and K columns:
\describe{
  \item{<col1>}{<short description>}
  \item{<col2>}{<short description>}
  \item{<col3>}{<short description>}
  % ... add one \item per column based on names(fire)
}
}
\source{
Local file \code{FIREDAM.csv} provided in course materials.
}
\usage{
fire
}
\description{
A dataset containing fire- and damage-related variables used for examples in MATH4753.
}
\examples{
data("fire")
head(fire)
}
\keyword{datasets}
