% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab2.R
\name{lab2}
\alias{lab2}
\title{Z score from a vector
Compute Z score for outlier detection
Input parameter must be a vector}
\usage{
lab2(x)
}
\arguments{
\item{x}{a vector of interest}
}
\value{
z score of a vector
}
\description{
Z score from a vector
Compute Z score for outlier detection
Input parameter must be a vector
}
\examples{
x = df1$COLUMN1; lab2(x)
}
