% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab7.R
\name{lab7chisim2}
\alias{lab7chisim2}
\title{Two Population Chi Sq Statistics}
\usage{
lab7chisim2(
  n1 = 10,
  n2 = 14,
  sigma1 = 3,
  sigma2 = 3,
  mean1 = 5,
  mean2 = 10,
  iter = 1000,
  xleg = 12,
  ymax = 0.07,
  ...
)
}
\arguments{
\item{n1}{n sample of y1}

\item{n2}{n sample of y2}

\item{sigma1}{std dev of y1}

\item{sigma2}{std dev of y2}

\item{mean1}{mean of y1}

\item{mean2}{mean of y2}

\item{iter}{number of iteration}

\item{xleg}{enable user toggle legend by click on the graphic output}

\item{ymax}{max of y, for visual clearance}

\item{...}{}
}
\value{
Chi sq statistics graphics
}
\description{
Two Population Chi Sq Statistics
}
\examples{
lab7chisim2(n1=10,n2=14,sigma1=3,sigma2=3,mean1=5,mean2=10,iter=1000, ...)
}
