% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Title: Binomial Distribution Simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{iteration variable}

\item{n}{number of samples}

\item{p}{probability of success}
}
\value{
A barplot of the proportions of the successes
}
\description{
Title: Binomial Distribution Simulation
}
\examples{
mybin(iter=100,n=10, p=0.7)
}
