% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybinom.R
\name{mybin}
\alias{mybin}
\title{Binomial function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations to perform}

\item{n}{number of trials}

\item{p}{probability of success of each trial}
}
\value{
numeric vector
}
\description{
Binomial function
}
\examples{
mybin(iter=100,n=10, p=0.5)
}
