% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Makes a binomial distribution plot}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{sample size}

\item{p}{probability of success for each sample}
}
\value{
a binomial distribution plot
}
\description{
Makes a binomial distribution plot
}
\examples{
mybin(iter=1000,n=18, p=0.3)

}
