% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{My Binomial}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations to run}

\item{n}{Number of trials}

\item{p}{Probability of Success}
}
\value{
A table of probabilities for the # of samples along with outputting a barplot
}
\description{
My Binomial
}
\examples{
mybin(iter = 10, n = 10, p = 0.5)
}
