% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybinomial.R
\name{mybin}
\alias{mybin}
\title{Binomial Distribution}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Sample size}

\item{p}{Probability}
}
\value{
A binomial distribution barplot
}
\description{
Binomial Distribution
}
