% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5.R
\name{mybin}
\alias{mybin}
\title{The Binomial Simulation
Creates bin distribution for binomial}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{number of bins}

\item{p}{probablity of success}
}
\value{
binomial simulation graphics
}
\description{
The Binomial Simulation
Creates bin distribution for binomial
}
\examples{
mybin(iter = 25, n = 9, p = 075)
}
