% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5.R
\name{mybin}
\alias{mybin}
\title{Mybin function for lab 5}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{p}{}
}
\value{

}
\description{
Mybin function for lab 5
}
\examples{
mybin(iter = 25, n = 9, p = 075)
}
