% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{mybin}
\alias{mybin}
\title{Binomial}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Iterations}

\item{n}{Size}

\item{p}{Probability}
}
\value{
A bar graph simulating the binomial distribution
given some sample size and probability of success
}
\description{
Binomial
}
\examples{
\dontrun{mybin)()}
}
