% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{my bin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{iterations}

\item{n}{sample size}

\item{p}{probability}
}
\value{
binomial experiment with plot and table
}
\description{
my bin
}
\examples{
\dontrun{mybin(iter=1000,n=18, p=0.3)}
}
