% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Simulate Binomial Distribution and Plot}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Number of trials}

\item{p}{Probability of success}
}
\value{
Relative frequencies as a table
}
\description{
This function simulates a binomial distribution using rbinom and
produces a barplot of relative frequencies.
}
