% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Title}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{Sample size}

\item{p}{probability of success}
}
\value{
A table of proportions
}
\description{
Title
}
\examples{
mybin(iter = 100, n = 10, p = 0.5)
}
