% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{myBinomial}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{How many iterations to run through}

\item{n}{Size of population}

\item{p}{Probability of a hit (p<=1)}
}
\value{
A table of the probability of a cetrain amount of hits occuring
}
\description{
myBinomial
}
\examples{
mybin(100,10,0.5)
}
