% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Mybin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{the number of iterations}

\item{n}{the number of Bernoulli trials}

\item{p}{the probability of success in each trial}
}
\value{
barplot of relative frequencies and  a table of the same
}
\description{
Mybin
}
\examples{
mybin()
}
