% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{My binary distribution generator}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{a number of simulations to run}

\item{n}{a number of decisions to make in each iterations}

\item{p}{a number probability that each decision is a success}
}
\value{
a table with the probability for each number of success in 0:n
}
\description{
My binary distribution generator
}
\examples{
probTable<-mybin(iter=10000,n=10,p=0.7)
}
