% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBinomial_function.R, R/myZfunction.R
\name{mybin}
\alias{mybin}
\title{myBinomial
calculates the binomial distribution}
\usage{
mybin(iter = 100, n = 10, p = 0.5)

mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{an integer}

\item{n}{an integer}

\item{p}{a double}
}
\value{
binomial distribution, barplot and table of distributions

binomial distribution, barplot and table of distributions
}
\description{
myBinomial
calculates the binomial distribution

myBinomial
calculates the binomial distribution
}
\examples{
iter = 10, n = 3, p = 0.5; mybin(iter,n,p);
iter = 10, n = 3, p = 0.5; mybin(iter,n,p);
}
