% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{The number of iterations.}

\item{n}{The number of trials.}

\item{p}{The probability (0<=p<=1).}
}
\value{
A bar plot of the successes and a table showing the chances to get a set amount of successes.
}
\description{
mybin
}
\examples{
mybin(100, 5, .5)
}
