% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{How many iterations}

\item{n}{number of values}

\item{p}{probability of success}
}
\value{
A barplot of sucesses from a simulated binomial function.
}
\description{
mybin
}
\examples{
\dontrun{mybin(iter=1000,n=18, p=0.3)}
}
