% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{binomal simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
a barplot of the proportion of successes
}
\description{
binomal simulation
}
\examples{
\dontrun{mybin(iter=i, n=10, p=0.7)}
}
