% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBin.R
\name{myBin}
\alias{myBin}
\title{Title Binomial Simulation}
\usage{
myBin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{An integer specifying the number of iterations (or experiments) to run.}

\item{n}{An integer specifying the number of trials per iteration.}

\item{p}{A numeric value between 0 and 1 specifying the probability of success for each trial.}
}
\value{
A named vector of proportions showing the frequency of the number of successes
out of `n` trials over all `iter` iterations.
}
\description{
Title Binomial Simulation
}
\examples{
myBin(iter = 100, n = 10, p = 0.7)

}
