% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin1.R
\name{mybin}
\alias{mybin}
\title{mybin1}
\usage{
mybin(iter = 1000, n = 10, p = 0.5)
}
\arguments{
\item{iter}{a number}

\item{n}{a number}

\item{p}{a number}
}
\value{
a barplot
}
\description{
mybin1
}
\examples{
mybin(iter = 100, n = 10, p = 0.7)
}
