% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{mybin}
\alias{mybin}
\title{binomial function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{number of samples}

\item{p}{probability}
}
\value{
binomial simulation
}
\description{
binomial function
}
\examples{
\dontrun{mybin(iter=100,n=10, p=0.5)}
}
