% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyBinom.R
\name{mybin}
\alias{mybin}
\title{Binomial Sampling Function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Number of trials}

\item{p}{probability that event A will occur/ a success will occur}
}
\value{
plot of the simulation of and a table of proportions
}
\description{
This uses the sample function to produce a binomial simulation and plots the simulation and outputs a table of proportions
}
\examples{
mybin(iter = 100, n = 10, p = 0.5)
mybin()
}
