% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Standard bootstrap function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{the number of iterations of sampling desired. i.e. iter =100, 100 samples collected}

\item{x}{it a vector of random variables to be studied}

\item{fun}{the function that is to be applied to the materix of n*iter}

\item{alpha}{1-alpha is the confidence interval desired. i.e. if 95\% interval desired, alpha is 0.05}

\item{cx}{size of the data points on the graph}

\item{...}{allows for more arguments to be put in the function}
}
\value{
returns a histogram for a chosen statistic, i.e. mean, chi-squared
}
\description{
Standard bootstrap function
}
\examples{
  \dontrun{myboot2(x, fun="sd", alpha =0.3)}
}
