% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Confidence Interval Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Bootstrap iterations}

\item{x}{Data vector}

\item{fun}{Statistic}

\item{alpha}{Alpha level}

\item{cx}{Text size}

\item{...}{Additional parameters}
}
\value{
A list with the confidence interval, function used, and data.
}
\description{
Creates a bootstrap confidence interval
}
