% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{My bootstrap function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterations}

\item{x}{sample size}

\item{fun}{type of function}

\item{alpha}{alpha}

\item{cx}{numerical value to make elements larger}

\item{...}{more parameters}
}
\value{
Bootstrap intervals
}
\description{
My bootstrap function
}
\examples{
myboot2(iter = 10000, x = 20, fun = "mean", alpha = 0.05, cex = 1.5)

}
