% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{My Bootstrap2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations to run}

\item{x}{the vector of samples}

\item{fun}{the function to use}

\item{alpha}{100*1-alpha \% c.i.}

\item{cx}{relative magnification of text and symbols}

\item{...}{additional histogram options}
}
\value{
Histogram of results after "fun" is applied with a (100-alpha)\% confidence interval labeled and a point estimate.
}
\description{
My Bootstrap2
}
\examples{
\dontrun{myboot2(iter = 500, x = sample, fun = "mean", alpha = 0.05)}
}
