% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2
a function that takes in data and with that data makes a specified number of sample statistics from the population.
These statistics can then be used to make confidence intervals of the entire population.}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{- how many samples you want}

\item{x}{- the data}

\item{fun}{- the sample statistic that you want to find}

\item{alpha}{- alpha represents the confidence interval. 1- aplha is the % confidence you want}

\item{cx}{- defines how much the text on the graph should be scaled}

\item{...}{- a parameter that lets you set your own parameter for the histogram}
}
\value{
a histogram of the distribution of the samples which has a histogram, the confidence interval, each samples statistic, the function used
}
\description{
myboot2
a function that takes in data and with that data makes a specified number of sample statistics from the population.
These statistics can then be used to make confidence intervals of the entire population.
}
\examples{
myboot2(10000,x=sam,fun="mean",alpha=0.05,xlab="mean",col="Purple",cx=1.5)
}
