% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBoot2.R
\name{myBoot2}
\alias{myBoot2}
\title{BootStrap Function}
\usage{
myBoot2(iter = 10000, x, fun = "var", alpha = 0.2, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{size of the sample}

\item{fun}{name of the function}

\item{alpha}{1 - % , number of alpha}

\item{cx}{size of the wording}

\item{...}{more parameters avaliable

return list of data if wanted
returns a barplot by default}
}
\value{

}
\description{
A function that takes in a number of paramters and simulates a bootstrap barplot
}
\examples{
myBoot2(iter = 10000, x = 20, fun = "mean", alpha = 0.05, cx = 1)
}
