% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBoot2.R
\name{myBoot2}
\alias{myBoot2}
\title{BootStrap Function}
\usage{
myBoot2(iter = 10000, x, fun = "var", alpha = 0.2, cx = 1.5, ...)
}
\arguments{
\item{iter}{size of the sample}

\item{x}{name of the function}

\item{fun}{1 - % , number of alpha}

\item{alpha}{size of the wording}

\item{cx}{more parameters avaliable}

\item{...}{return list of data if wanted
returns a barplot by default}
}
\value{

}
\description{
number of iterations
}
\examples{
myBoot2(iter = 10000, x = 20, fun = "mean", alpha = 0.05, cx = 1)
}
