% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab9.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap simulation for confidence interval}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iteration}

\item{x}{sample to be bootstrapped}

\item{fun}{function of interest to be estimated}

\item{alpha}{alpha parameter of confidence interval}

\item{cx}{point graphical variable for visual modification}

\item{...}{}
}
\value{
1-alpha CI for a particular function
}
\description{
predict a function output with (1-alpha) confidence interval from a sample.
default value evaluates the input with 10,000 iterations to estimate the mean
}
\examples{
myboot2(x = rnorm(20, mean = 10, sd = 2), alpha = 0.1, fun = "sd")
}
