% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap estimation}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{the number of samples created from the data set}

\item{x}{the data set of interest}

\item{fun}{The parameter/function of interest. Mean by default}

\item{alpha}{The parameter used to find the confidence interval of interest}

\item{cx}{The size of the plotting text relative the base value of 1}

\item{...}{Any additional comments to be added to the title}
}
\value{
A histogram of Bootstrap sample statistics
}
\description{
Bootstrap estimation
}
\examples{
\dontrun{myboot2<-function(iter=10000,x,fun="mean",alpha=0.05,cx=1.5,...)}
}
