% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, xName = "x")
}
\arguments{
\item{iter}{number of iterations}

\item{x}{the data}

\item{fun}{function to use}

\item{alpha}{aloha}

\item{cx}{cx}

\item{xName}{Name for x on the histogram}
}
\value{
a histogram and a list of important stuff
}
\description{
myboot2
}
\examples{
\dontrun{myboot2(x=sam,fun="median", xName = "median")}
}
