% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrat confidence interval}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterations}

\item{x}{sample}

\item{fun}{function}

\item{alpha}{alpha}

\item{cx}{cx}

\item{...}{}
}
\value{
a plot with confidence interval
}
\description{
Bootstrat confidence interval
}
\examples{
myboot2(x=c(1,2))
}
