% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBoot.R
\name{myboot2}
\alias{myboot2}
\title{MyBoot}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{...}{}
}
\value{
a histrogram using the bootstrap function
}
\description{
MyBoot
}
\examples{
set.seed(39); sam=rnorm(25,mean=25,sd=10)
windows();myboot2(10000,x=sam,fun="mean",alpha=0.05,xlab="mean",col="Purple",cx=1.5) # mac quartz()
}
