% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myBoot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{The amount of iterations of sampling}

\item{x}{The data set to draw samples from}

\item{fun}{The function to apply to the samples}

\item{alpha}{Confidence interval alpha}

\item{cx}{symbol scale}

\item{...}{any additional parameters}
}
\value{
A histogram of the sample distributions given the function
}
\description{
myBoot2
}
\examples{
myboot2(x=c(1:100))
}
