% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Intervals}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of bootstrap iterations to perform}

\item{x}{The original sample data}

\item{fun}{The statistic to compute e.g., mean, median}

\item{alpha}{Significance level for the confidence interval}

\item{cx}{Text size scaling factor for labels on the plot}

\item{...}{Additional arguments passed to the hist function}
}
\value{
A list containing confidence interval, function, and sample data
}
\description{
Bootstrap Intervals
}
\examples{
x <- rnorm(30)
myboot2(x = x, fun = "mean", iter = 5000)

}
