% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{My bootstrap program}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{the number of samples to take for a bootstrap simulation, default 10000}

\item{x}{a vector containing the population}

\item{fun}{a character string with the name of the function to be examined, default is "mean"}

\item{alpha}{the parameter determining the confidence level of the interval to 1-alpha percent, default is 0.05}

\item{cx}{a character expansion factor applied to the text added to the histogram}

\item{...}{a series of parameters passed to the hist() call that generates the histogram}
}
\value{
ci, the confidence interval

x, the input population

fun, the function being examined
}
\description{
Runs a bootstrap simulation to create a confidence interval
}
\details{
Given a vector containing the population, creates iter random samples and runs a bootstrap to create a confidence interval of (1-alpha)%. Also produces a histogram of the resulting sampling distribution.
}
\examples{

set.seed(20)
sam=rnorm(20,mean=10,sd=4)
myboot2(x = sam)
}
