% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{My Boot 2 Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{times to sample}

\item{x}{distribution}

\item{fun}{describes what function to apply}

\item{alpha}{percentage}

\item{cx}{text scaling}

\item{...}{histogram properties}
}
\value{
Returns a histogram and list of data values used
}
\description{
Create a histogram and return the list of data values used to create said histogram
}
\examples{
obj=math4753::myboot2(x=fire$DAMAGE) (prints a histogram and puts data values into obj)
}
