% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{My bootstrap function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of times to sample the provided data}

\item{x}{Sample data provided}

\item{fun}{Function to be applied to sampled data}

\item{alpha}{Confidence interval, given as the complement of the desired interval}

\item{cx}{Scaling of text for plots}

\item{...}{Possible additional parameters}
}
\value{
A histogram of the desired function
}
\description{
My bootstrap function
}
\examples{
 \dontrun{myboot(10000,x=sam,fun=function(x) mean(x)/median(x),alpha=0.05,xlab="mean(x)/median(x)",col="Blue")}
}
