% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R, R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myBoot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)

myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{an integer (iterations)}

\item{x}{a vector (sample)}

\item{fun}{a string (the function you want to perform like mean, median etc)}

\item{alpha}{an int, confidence interval}

\item{...}{}
}
\value{
a barplot that shows given confidence interval

a barplot that shows given confidence interval
}
\description{
myBoot2

myBoot2
}
\examples{
myboot2(iter=10000,x = sample,fun="median",alpha=0.09,cx=1.5)
myboot2(iter=10000,x = sample,fun="median",alpha=0.09,cx=1.5)
}
