% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{bootstrap function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of bootstrap iterations (default: 10000)}

\item{x}{Numeric vector of data points}

\item{fun}{Function to calculate the statistic (default: mean)}

\item{alpha}{Significance level for confidence interval (default: 0.05)}

\item{cx}{Control parameter for text size in plots.}

\item{...}{Additional arguments passed to other functions, such as graphical parameters for `hist`.}
}
\value{
A list containing the confidence interval, statistic function, original data, and point estimate
}
\description{
bootstrap function
}
\examples{
\dontrun{data <- rnorm(100) result <- myboot2(x=data, iter=1000, fun=mean)}

}
