% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{course 4753}
\usage{
myboot2(
  iter = 10000,
  x,
  fun = "mean",
  alpha = 0.05,
  xmax = 1000,
  ymax = 1000,
  ...
)
}
\arguments{
\item{x}{to find the condident interval mean and make a plot}
}
\value{
a histogram plot with mean and confident interval
}
\description{
to find the condident interval, mean and make a plot
}
\examples{
myboot2(x = ddt$LENGTH)


}
