% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Title Bootstrap Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{data vector}

\item{fun}{function to apply}

\item{alpha}{1 - confidence level}

\item{cx}{character expansion for text}

\item{...}{}
}
\value{
a list with confidence interval, function used, and data
}
\description{
Title Bootstrap Function
}
\examples{
\dontrun{myboot2(iter=10000,x=sam,fun=mean,alpha=0.05)}
}
