% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab9func.R
\name{myboot2}
\alias{myboot2}
\title{Title}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterations}

\item{x}{sample data}

\item{fun}{mean}

\item{alpha}{alpha}

\item{cx}{cx}

\item{...}{...}
}
\value{
interval
}
\description{
Title
}
\examples{
\dontrun{myboot2(sam, iter = 10000, alpha = 0.05)}
}
