% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot2}
\alias{myboot2}
\title{BootStrap Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{...}{}
}
\value{
return list of data if wanted
returns a barplot by default
}
\description{
A function that takes in a number of paramters and simulates a bootstrap barplot
}
\examples{
myBoot2(iter = 10000, x = 20, fun = "mean", alpha = 0.05, cx = 1)
}
