% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab7func.R
\name{mychisim}
\alias{mychisim}
\title{mychisim RMD version from Lab 7}
\usage{
mychisim(n1 = 10, sigma1 = 3, mean1 = 5, iter = 1000, ymax = 0.1, ...)
}
\arguments{
\item{n1}{sample size}

\item{sigma1}{standard deviation}

\item{mean1}{mean}

\item{iter}{iterations}

\item{ymax}{Y max on the graph}

\item{x}{vector x}

\item{y}{vector y}
}
\value{
chi plot
}
\description{
mychisim RMD version from Lab 7
}
\examples{
\dontrun{mychisim(iter = 10000, ymax = 0.15)}
}
