% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Computes a 95% confidence interval for the mean}
\usage{
myci(x)
}
\arguments{
\item{x}{Vector representing the sample data}
}
\value{
Vector of length 2, for lower and upper bounds of the confidence interval
}
\description{
Computes a 95% confidence interval for the mean
}
\examples{
set.seed(23)
x <- rnorm(30, mean = 10, sd = 12)
myci(x)

}
