% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{myci}
\usage{
myci(x)
}
\arguments{
\item{x}{sample of the population}
}
\value{
95% confidence interval of the population mean
}
\description{
given a sample of a population, my ci creates a 95% confidence interval of the population mean
}
\examples{
myci(x)
}
