% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{95% confidence interval generating function}
\usage{
myci(x)
}
\arguments{
\item{x}{a vector of samplings}
}
\value{
confidence interval for the given vector
}
\description{
A function that will take in a vector of samples and return a 95% confidence interval for the given data
}
\examples{
x = rnorm(30, mean = 10, sd = 12)
myci(x)
}
