% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Title}
\usage{
myci(x, n = 30, alpha = 0.05/2)
}
\arguments{
\item{n}{sample size}

\item{alpha}{confidence %}
}
\value{
confidence interval
}
\description{
Title
}
\examples{
myci(x)
}
