% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab11.R
\name{myci}
\alias{myci}
\title{95% Confidence Inteval}
\usage{
myci(x)
}
\arguments{
\item{x}{a vector of interest}
}
\value{
a 95% lower and upper confidence interval
}
\description{
Create a 95% confidence interval from a given vector
}
\examples{
myci(x)
}
