% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{My Confidence Intervals}
\usage{
myci(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
This uses the T test to show the confidence intervals for a 95% confidence level.
}
\examples{
set.seed(23);x = rnorm(30,mean=10,sd=12)
math4753::myci(x)
}
