% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{95% Confidence Interval Function for the Mean of a Sample}
\usage{
myci(x)
}
\arguments{
\item{x}{the vector of the sample we choose to find the confidence interval for}
}
\value{
vector with an interval for the 95% confidence level
}
\description{
Finds the 95% confidence interval for a sample with mean mu
}
\details{
This is the 4th function in this package. This function produces a vector of an interval. This interval shows the confidence interval for the 95% confidence level of the mean of a sample which is being processed or which goes through this function.
}
\examples{
set.seed(23)
x = rnorm(30,mean=10,sd=12)
myci(x)
}
