% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{myci}
\usage{
myci(sample = x)
}
\arguments{
\item{t.test}{}
}
\value{
confidence interval for the sample at the confidence level of 95%
}
\description{
myci
}
\examples{
set.seed(23);x = rnorm(30,mean=10,sd=12)

}
