% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{course 4753}
\usage{
myci(x)
}
\arguments{
\item{x}{to find confident interval.}
}
\value{
two values for confident interval.
}
\description{
Use rnorm to find confident interval
}
\examples{
x = rnorm(30,mean=10,sd=12)
myci(x)



}
