% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{95% confidence interval generating function
Confident interval}
\usage{
myci(x)
}
\arguments{
\item{x}{}
}
\value{
confidence interval for the given vector
}
\description{
a function that will create a 95% confidence interval for mean x from a single sample x
}
\examples{
x = rnorm(30, mean = 10, sd = 12)
myci(x)
}
