% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{myci}
\usage{
myci(x, alpha = 0.05)
}
\arguments{
\item{x}{an array containing a random sample of a population}

\item{alpha}{the probability of a Type 1 error used to create the confidence level}
}
\value{
A list containing (lower bound, upper bound) of the CI for the population mean with confidence level (1-alpha)100%
}
\description{
myci
}
\examples{
myci(x = rnorm(n=20, mean = 5, sd = 2), alpha = .05)
}
