% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Central Limit Theorem Function}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{The sample size}

\item{iter}{The number of iterations}

\item{a}{The lower limit}

\item{b}{The upper limit}
}
\value{
A list of sums for each group of the numbers.
}
\description{
This function creates random numbers from a uniform distribution then adds them to create a histogram of sums.
}
\examples{
w <- myclt(n=50, iter=10000)
}
