% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{My Central Limit Function}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{the number of times to repeat the iter process (the mean of this many will be taken)}

\item{iter}{the number of times a data value will be taken}

\item{a}{the lowest value to be sampled}

\item{b}{the highest value to be sampled}
}
\value{
a vector of means (length determined by iter)
}
\description{
Creates a Central limit vector that is put into a histogram.
}
\examples{
obj=math4753::myclt(n=5,iter=100000) (puts a central limit vector into obj and prints a histogram)
}
