% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltr.R
\name{myclt}
\alias{myclt}
\title{Scatter histogram}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{the number of samples per iteration}

\item{iter}{the number of iterations to perform}

\item{a}{Lower limit for uniform distribution (default is 0).}

\item{b}{Upper limit for uniform distribution (default is 5).}
}
\value{
Invisible list containing the means of samples
}
\description{
Scatter histogram
}
\examples{
\dontrun{w=myclt(n=50,iter=10000,a=5,b=10)}

}
