% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Central Limit Theorem function}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}
}
\value{
a graph and a vector of the sum of the uniform random variables
}
\description{
Central Limit Theorem function
}
\examples{
\dontrun{myclt(n=10,iter=10000)}
}
