% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab8func.R
\name{myclt}
\alias{myclt}
\title{CLT: my Central Limit Function}
\usage{
myclt(n = 50, iter, a = 0, b = 5)
}
\arguments{
\item{n}{sample size}

\item{iter}{iterations}

\item{a}{lower limit}

\item{b}{upper limit}
}
\value{
CLT function
}
\description{
CLT: my Central Limit Function
}
\examples{
\dontrun{myclt(n=50,iter=10000,a=5,b=10)}
}
