% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{mycltb}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{the sample size to be observed}

\item{iter}{number of iterations over a sample size}

\item{p}{the probability of a success in a trial}

\item{...}{allows user to put in optional arguments}
}
\value{
returns a histogram of the means of sample size n for a binomial distribution
}
\description{
mycltb
}
\examples{
   \dontrun{mycltb(n=5, iter=10000, p = 0.7)}
}
