% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{My CLT Binomial Simulation}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of samples}

\item{p}{probability for binomial}

\item{...}{various other arguments}
}
\value{
histogram of the iter samples of size n
}
\description{
My CLT Binomial Simulation
}
\examples{
mycltb(n = 10, iter = 100, p = 0.4)
}
