% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltb}
\alias{mycltb}
\title{Histogram of Sample Mean from Binomial Distribution}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{Numeric. Sample size for each iteration.}

\item{iter}{Numeric. Number of iterations or samples to be taken.}

\item{p}{Numeric. Probability of success for the binomial distribution. Default is 0.5.}

\item{...}{Additional parameters to be passed to the `hist` function.}
}
\value{
Invisible. A histogram is plotted as a side effect.
}
\description{
This function simulates samples from a binomial distribution, computes their means,
and plots a histogram of these means with a theoretical normal curve overlay.
}
\examples{
\dontrun{
mycltb(n=50, iter=1000)
}
}
