% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{mtcltb- This function uses a binomial distribution with probability p}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{Sample size}

\item{iter}{Number of iterations}

\item{p}{Probability of success on each trial}

\item{...}{Possible additional parameters}
}
\value{
Creates a histogram showing the results
}
\description{
mtcltb- This function uses a binomial distribution with probability p
}
\examples{
mycltb(n=4,iter=10000,p=0.5)
}
