% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab8func.R
\name{mycltb}
\alias{mycltb}
\title{CLT Binomial: my Central Limit Function BINOMIAL. CLT will work with discrete or continuous distributions}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{iterations}

\item{p}{probability}

\item{...}{...}
}
\value{
CLT binomial function
}
\description{
CLT Binomial: my Central Limit Function BINOMIAL. CLT will work with discrete or continuous distributions
}
\examples{
\dontrun{mycltb(n=5,iter=10000,p=0.5)}
}
