% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{CLT Binomial
 A function that takes in the number of sample, n and p, and simulates a Binomial histogram and plot}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{...}{}
}
\value{
three graphs including probability, density and rel. frequency
}
\description{
CLT Binomial
 A function that takes in the number of sample, n and p, and simulates a Binomial histogram and plot
}
\examples{
mycltp(100,123,10)
}
