% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp
This function plots the mean distributions of a poisson distribution. And it makes a histogram of the sample mean
It also makes a barplot of the probability of a value being chosen.}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n-}{number of samples per iteration}

\item{iterations-}{number of iterations}

\item{lambda-}{lambda of a poisson distribution}
}
\value{

}
\description{
mycltp
This function plots the mean distributions of a poisson distribution. And it makes a histogram of the sample mean
It also makes a barplot of the probability of a value being chosen.
}
\examples{
mycltp(n=2, iter=10000, lambda=10)
}
