% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{CLT Poisson}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{number of iterations}

\item{iter}{number of lambda}

\item{lambda}{any other parameters that may be added}

\item{...}{}
}
\value{
three graphs including probability, density and rel. frequency
}
\description{
number of sample
}
\examples{
mycltp(100,10000,10)
}
