% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{CLT Poisson}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{number of sample}

\item{iter}{number of iterations}

\item{lambda}{number of lambda}

\item{...}{any other parameters that may be added}
}
\value{
three graphs including probability, density and rel. frequency
}
\description{
A function that takes in the number of sample, iteration and lambda, and simulates a Poisson histogram and plot
}
\examples{
mycltp(100,10000,10)
}
