% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cltpoisson.R
\name{mycltp}
\alias{mycltp}
\title{Central Limit Theorem for Poisson}
\usage{
mycltp(n = 10, iter = 10000, lambda = 10, ...)
}
\arguments{
\item{n}{The number of observation in a sample}

\item{iter}{The number of random samples}

\item{lambda}{The parameter of the poisson}

\item{...}{A dynamic parameter that can take the value of any specified parameter.}
}
\value{
A histogram showing the density of sample means taken
from iter number of samples with n number of observations
per sample. A barplot with the relative frequncies of y
values  vs the value A plot of probability of the y value
vs the y value.
}
\description{
Central Limit Theorem for Poisson
}
\examples{
\dontrun{mycltp(n=10,iter=10000, lambda = 10)}
}
