% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Poisson central limit}
\usage{
mycltp(n, iter, lambda = 10)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{lambda}{lamda}
}
\value{
Graphs having to do with the central limit theorem and the Poisson distribution.
}
\description{
Poisson central limit
}
\examples{
\dontrun{mycltp(n=20,iter=10000,lambda=4)}
}
