% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{My Poisson Central Limit Theorem}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{the number of elements per sample}

\item{iter}{the number of samples to take}

\item{lambda}{the mean and variation of the source Poisson distribution}

\item{...}{futher arguments passed to the call of hist}
}
\value{

}
\description{
Reports the distribution of sample means for iter samples of size n.
}
\details{
Given the characteristic number for the Poisson distribution, takes iter samples of size n, then calculates the mean for each sample and reports the distribution.
}
\examples{

mycltp(n=3, iter = 1000, lambda = 5)
}
