% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab8func.R
\name{mycltp}
\alias{mycltp}
\title{CLT Poisson: my Central Limit Function POISSON. Will work with discrete or continuous distributions.}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{iterations}

\item{lambda}{lambda: average rate of occurrence}

\item{...}{...}
}
\value{
CLT Poisson function
}
\description{
CLT Poisson: my Central Limit Function POISSON. Will work with discrete or continuous distributions.
}
\examples{
\dontrun{mycltp(n=10,iter=10000)}
}
