% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Simulate sampling distribution of Poisson sample means}
\usage{
mycltp(n, iter, lambda)
}
\arguments{
\item{n}{Sample size}

\item{iter}{Number of iterations}

\item{lambda}{Poisson mean}
}
\value{
Numeric vector of sample means (and histogram)
}
\description{
Simulate sampling distribution of Poisson sample means
}
