% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Central limit Uniform function}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{sample size}

\item{iter}{iterations}

\item{a}{lower limit}

\item{b}{upper limit}
}
\value{
Histogram of sample mean with theoretical curves
}
\description{
Central limit Uniform function
}
\examples{
 mycltu(20, 10000)

}
