% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{My CLT - Uniform Distribution}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{the sample size}

\item{iter}{the number of samples run}

\item{a}{lower limit for uniform dist}

\item{b}{upper limit for uniform dist}
}
\value{
a plot with a histogram of sample means, the expected density curve (red), the actual density curve (blue), and the density from which the samples were taken (black)
}
\description{
A function showing an application of the central limit theorem
on a uniformly distributed random variable
}
\examples{
mycltu(n=10,iter=10000)
mycltu(n=5,iter=20000,a=1,b=7)
}
