% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltpu.R
\name{mycltu}
\alias{mycltu}
\title{Title mycltu}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{a}{lower limit}

\item{b}{upper limit}
}
\value{
histogram with information
}
\description{
Title mycltu
}
\examples{
\dontrun{mycltu(n=54,iter=100000)}
}
