% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Uniform distribution CLT}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{sample size}

\item{iter}{iterations}

\item{a}{lower bound}

\item{b}{upper bound}
}
\value{
a plot
}
\description{
Uniform distribution CLT
}
\examples{
mycltu(10,10000)
}
