% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{mycltu}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{The size of each sample created}

\item{iter}{The number of samples generated}

\item{a}{The minimum value}

\item{b}{The maximum value}
}
\value{
A density curve and a histogram of the samples.
}
\description{
mycltu
}
\examples{
mycltu(5, 10000, 0, 10)
}
