% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab8func.R
\name{mycltu}
\alias{mycltu}
\title{CLT uniform: my Central Limit Function}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{sample size}

\item{iter}{iterations}

\item{a}{lower limit}

\item{b}{upper limit}
}
\value{
CLT uniform function
}
\description{
CLT uniform: my Central Limit Function
}
\examples{
\dontrun{mycltu(n = 10, iter = 10000, a = 0, b = 10)}
}
