% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myf.R
\name{myf}
\alias{myf}
\title{My function of sqaures and cubes}
\usage{
myf(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a list of vectors
}
\description{
produce list of squares and cubes
}
\details{
This is a part of the introduction to package making
}
\examples{
x=1:4;myf(x)
}
