% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Square function}
\usage{
myf(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
a vector of squared components in the form of vector
}
\description{
Takes a vector and returns a vector of squared components
}
\examples{
x=1:30; myf(x)
}
