% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Title
Make a list of sqares and cubes
Intro to package making stuff}
\usage{
myf(x)
}
\arguments{
\item{x}{}
}
\value{
a list of vector
}
\description{
Title
Make a list of sqares and cubes
Intro to package making stuff
}
\examples{
x = 1:4; myf(x)
}
