% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{My quadratic function}
\usage{
myf(x)
}
\arguments{
\item{x}{a vector of double values}
}
\value{
a vector of double values
}
\description{
Creates a quadratic function on a vector
}
\details{
This is a test function for working on packages
}
\examples{
x = 1:30;
myf(x);

[1]    3    4    9   18   31   48   69   94  123  156  193  234  279
[14]  328  381  438  499  564  633  706  783  864  949 1038 1131 1228
[27] 1329 1434 1543 1656
}
