% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Cube function}
\usage{
myf(x)
}
\arguments{
\item{x}{}
}
\value{
a vector of cubed components
}
\description{
Takes a vector, cubes it, and returns the new vector
}
\examples{
x=1:30; myf(x)
}
