% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{myf}
\alias{myf}
\title{My quadratic}
\usage{
myf(x)
}
\arguments{
\item{x}{is a vector of double values}
}
\value{
a vector of doubles
}
\description{
Creates a quadratic on a vector
}
\details{
This is a test function for working on packages
}
\examples{
x = 1:30;myf(x)
}
