% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{My quadratic function}
\usage{
myf(x)
}
\arguments{
\item{x}{a vector of double values}
}
\value{
a vector of double values
}
\description{
Creates a quadtratic on a vector
}
\details{
This is a test function for working on packages.
}
\examples{
x=1:30; myf(x)
}
