% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{myfun}
\alias{myfun}
\title{My First Function}
\usage{
myfun(x)
}
\arguments{
\item{x}{A Quantitative Vector}
}
\value{
A Quantitative Vector, the Square of the input
}
\description{
My First Function
}
\examples{
myfun(x = 1:10)
}
