% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{myfun}
\alias{myfun}
\title{Plot and Shade a Normal Distribution Curve}
\usage{
myfun(x)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sigma}{Standard deviation of the normal distribution.}

\item{a}{Value up to which area under the curve is shaded.}
}
\value{
A list containing the probability \(P(X \leq a)\).
}
\description{
Plot and Shade a Normal Distribution Curve
}
\examples{
\dontrun{
  result <- myncurve(0, 1, 0.5)
  print(result)
}
}
