% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{Hypergeometric simulation function}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{Number of iterations to perform}

\item{N}{number of items in the population}

\item{r}{number of items classified as a "1"}

\item{n}{number of items selected}
}
\value{
numeric vector with a barplot
}
\description{
Hypergeometric simulation function
}
\examples{
myhyper(iter=100,N=20,r=12,n=5)
}
