% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{Hypergeometric Simulation}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{number of iterations}

\item{N}{total number of elements}

\item{r}{number of successes in N}

\item{n}{number of elements drawn}
}
\value{
a plot
}
\description{
Hypergeometric Simulation
}
\examples{
 myhyper(iter=100,n=5,N=20,r=12)
}
