% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{My max likelihood function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{the likelihood function}

\item{x}{observed outcomes}

\item{param}{vector of probabilities to test}

\item{...}{extra parameters to pass to graphics functions}
}
\value{
a plot and a list of parameters
}
\description{
My max likelihood function
}
\examples{
mymaxlik(c(3,3,4,3,4,5,5,4),param=seq(0,1,length=10001),lfun=function(x,param){log(dbinom(x,prob=param,size=20))},xlab=expression(pi),main="Binomial",cex.main=2)
}
