% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{...}{}
}
\value{
the maximum likelihood estimate for the given sample
}
\description{
mymaxlik
}
\examples{
logpoiss=function(x,param) log(dpois(x,lambda=param))
}
