% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Estimation Plotter}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{A log-likelihood function taking (x, param).}

\item{x}{A numeric vector of observations.}

\item{param}{A numeric vector of parameter values to evaluate.}

\item{...}{Additional arguments to pass to the plot.}
}
\value{
A list containing:
\describe{
  \item{i}{Index of the maximum likelihood estimate.}
  \item{parami}{Parameter value corresponding to the maximum likelihood.}
  \item{yi}{Maximum log-likelihood value.}
  \item{slope}{Slope approximations around the maximum (if computable).}
}
}
\description{
This function evaluates and plots the log-likelihood for a parameter over a grid.
It uses the provided likelihood function and data to identify the maximum.
}
\examples{
loglik_binom <- function(y, p) dbinom(y, size = 20, prob = p, log = TRUE)
y <- c(3, 3, 4, 3, 4, 5, 5, 4)
pvals <- seq(0.01, 0.5, length = 100)
mymaxlik(loglik_binom, y, pvals)
}
