% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum likelihood function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{Function for the type of distribution we want to find the MLE of}

\item{x}{Vector of sample values}

\item{param}{Parameters for the distribution}

\item{...}{Possible additional parameters}
}
\value{
Creates a plot of the MLE, and outputs data about it
}
\description{
Maximum likelihood function
}
\examples{
\dontrun{mymaxlik(x=c(9,9,1,9,9,9),param=seq(0,1,length=1000),lfun=logbin,xlab=expression(pi),main="Binomial",cex.main=2)}
}
