% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMaxLik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{myMaxLik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{the function that you want to take ln of}

\item{x}{a vector of your trials}

\item{param}{a vectors, where the x values will be replaced to be in the ln format}

\item{...}{}
}
\value{

}
\description{
myMaxLik
}
\examples{
mymaxlik(x=c(4,6,7,6,5),param=seq(0,10,length=1000),lfun=logpoiss)
}
