% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMaxLike.R
\name{mymaxlik}
\alias{mymaxlik}
\title{myMaxLike}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{the funciton that will have ln applied}

\item{x}{a vector, of you trials}

\item{param}{a vector, which will have all of its values evaluated at ln}

\item{...}{}
}
\value{
a graph that will dislay the maximum likelyhood by looking at monotinicity
}
\description{
myMaxLike
}
\examples{
myMaxLike(x=c(3,3,4,3,4,5,5,4),param=seq(0,1,length=1000),lfun=logbin)
}
