% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function}

\item{x}{list of numbers}

\item{param}{sequence}

\item{...}{}
}
\value{
Finds the maximum likelihood estimate and records a plot.
}
\description{
mymaxlik
}
\examples{
\dontrun{mymaxlik(x = y, param = seq(0,1,length=1000), lfun = logbin, xlab = expression(pi),cex.main=1.0)}
}
