% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Estimation Function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{A likelihood function to be maximized.}

\item{x}{A vector of data points.}

\item{param}{A vector of parameter values to evaluate.}

\item{...}{Additional arguments passed to the plotting functions.}
}
\value{
A list containing the index of the maximum likelihood estimate, the parameter value at this index, the likelihood value, and slope analysis.
}
\description{
Calculates the maximum likelihood estimates for given parameters using a specified likelihood function and data.
}
\examples{
\dontrun{mymaxlik(x=c(9,9,1,9,9,9),param=seq(0,1,length=1000),
lfun=logbin,xlab=expression(pi),main="Binomial",cex.main=2)}
}
