% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab10func.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Title}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function}

\item{x}{data}

\item{param}{parameters}

\item{...}{...}
}
\value{
max likelihood
}
\description{
Title
}
\examples{
\dontrun{mymaxlik(x=c(9,9,1,9,9,9),param=seq(0,1,length=1000),lfun=logbin,xlab=expression(pi),main="Binomial",cex.main=2)}
}
