% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyMLE.R
\name{mymaxlik}
\alias{mymaxlik}
\title{My Maximum Likelihood Estimate}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{A log of the the joint probability density function}

\item{x}{The sample being passed in to create an MLE for a population parameter}

\item{param}{The parameter of the population that we are trying to estimate}

\item{...}{Whatever other paramters that can be added and will be included in the plot functions}
}
\value{
A plot containing the MLE on the graph and a list of text containing the x and y value of the maximum value
}
\description{
My Maximum Likelihood Estimate
}
\examples{
logbin=function(x,param) log(dbinom(x,prob=param,size=20));y = c(3,3,4,3,4,5,5,4);mymaxlik(lfun = logbin, x = y, param = seq(0,1, length = 1000), main = "Binomial, n = 20")
}
