% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlbeta.R
\name{mymlbeta}
\alias{mymlbeta}
\title{Maximum Likelihood Estimation for Beta Distribution}
\usage{
mymlbeta(x, alpha, beta, ...)
}
\arguments{
\item{x}{Number vector}

\item{alpha}{Number Vector}

\item{beta}{Number Vector}

\item{...}{Additional parameters}
}
\value{
A list with:
  \item{x}{Sample data}
  \item{coord}{Coordinates of the MLE}
  \item{maxl}{Maximum likelihood value}
  \item{maxalpha}{MLE for `alpha`.}
  \item{maxbeta}{MLE for `beta`.}
}
\description{
Maximum Likelihood Estimation for Beta Distribution
}
\examples{
# Example usage:
sample_data <- rbeta(30, shape1 = 3, shape2 = 4)
mymlbeta(sample_data, alpha = seq(1, 5, length = 100), beta = seq(2, 6, length = 100))

}
