% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlbeta.R
\name{mymlbeta}
\alias{mymlbeta}
\title{Title mymlbeta}
\usage{
mymlbeta(x, alpha, beta, ...)
}
\arguments{
\item{x}{- the sample data}

\item{alpha}{- the alpha value of the beta function}

\item{beta}{- the beta value of the beta function}

\item{...}{- a spot for an additional paramater when creating the contour graph}
}
\value{
the maximum likelihood estimations for alpha and beta
}
\description{
This function creates maximum likelihood estimations for a beta distribution. The function returns
a countour plot of all the maximum possibilities of the alpha and beta values.
}
