% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlbeta.R
\name{mymlbeta}
\alias{mymlbeta}
\title{mymlbeta}
\usage{
mymlbeta(x, alpha, beta, ...)
}
\arguments{
\item{x}{A vector representing a sample}

\item{alpha}{The alpha value in a beta distribution}

\item{beta}{The beta value in the beta distribution}

\item{...}{Extra variables that can change the contour that is generated}
}
\value{
A list that has the calculated values in the function
}
\description{
mymlbeta
}
\examples{
mymlbeta(x=rbeta(100,shape1=2,shape2=5),alpha=seq(1,4,length=100),beta=seq(2,8,length=100),lwd=2,labcex=1)
}
