% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlbeta.R
\name{mymlbeta}
\alias{mymlbeta}
\title{Title mlbeta function}
\usage{
mymlbeta(x, alpha, beta, ...)
}
\arguments{
\item{x}{sample vector}

\item{alpha}{vector of alpha values}

\item{beta}{vector of beta values}

\item{...}{additional graphical parameters}
}
\value{
A list containing the sample vector, coordinates of the maximum likelihood estimate, maximum likelihood value, and the estimated alpha and beta values.
}
\description{
Title mlbeta function
}
\examples{
\dontrun{mymlbeta(sample(rbeta(30, shape1 = 3, shape2 = 4) , length(rbeta(30, shape1 = 3, shape2 = 4) ), replace = TRUE), seq(1, 8, length.out = 100), seq(2, 10, length.out = 100))}
}
