% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlnormal.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Max Likelihood Estimator for Normal}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{the sample size}

\item{mu}{the values for the mean to be tested}

\item{sig}{the values for the standard deviation to be tested}

\item{...}{A filler parameter to add to the title of the plots, in this case}
}
\value{
A graphical estimate for both mu and sigma
}
\description{
Max Likelihood Estimator for Normal
}
\examples{
\dontrun{mymlnorm(x=c(10,12,13,15,12,11,10),mu=seq(10,15,length=1000),
sig=seq(0.1,4,length=1000),lwd=2,labcex=1)}
}
