% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbin.R
\name{mynbin}
\alias{mynbin}
\title{Title}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{y is the number of trials until the rth success is observed}

\item{r}{the number of successes}

\item{p}{the probability of success on a single bernoulli tria;}
}
\value{
Returns the negative Binomial probability of an event, i.e., the number of failures until the rth success
}
\description{
Title
}
\examples{
   \dontrun{mynbin(10,3,0.4)}
}
