% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbin.R
\name{mynbin}
\alias{mynbin}
\title{Sampling Function}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{p}{}
}
\value{

}
\description{
p = Probability of success on a single Bernoulli trial
r = 1 - p
y = Number of trials until the rth success is observed
}
\examples{
mybin(y,r,p)
}
