% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{#the value of x up to where the area under the graph should be colored}

\item{mu}{#the mean of the data}

\item{sigma}{#standard deviation of the data}
}
\value{
returns a plot of the normal distribution with the area/probability shaded in
        and written on the graph. Also lists the area into the command-line
}
\description{
myncurve
}
\examples{
   \dontrun{myncurve(10, 5, 6)}
}
