% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title: myncurve function implementation}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{a value}
}
\value{
A list of the mean and standard deviation
}
\description{
Title: myncurve function implementation
}
\examples{
myncurve(0, 1, 2)

}
