% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Shaded Normal Distribution Curve}
\usage{
myncurve(mu, sigma, a, b = -Inf, color = "blue")
}
\arguments{
\item{mu}{Mean of the distribution}

\item{sigma}{Standard deviation of the  distribution}

\item{a}{Upper limit for shaded area}

\item{b}{Lower limit for shaded area (default is negative infinity)}

\item{color}{Shaded area color}
}
\value{
A graph of the normal distribution curve and the probability of the shaded area
}
\description{
A function that returns a graph of the normal distribution curve and the probability of a specified shaded area
}
\examples{
# Example 1: Shading area under the curve from negative infinity to 1
myncurve(mu = 0, sigma = 1, a = 1)

# Example 2: Shading area under the curve from 1 to 2
myncurve(mu = 0, sigma = 1, a = 2, b = 1)

# Example 3: Shading area under the curve from -1 to 2 with a certain color
myncurve(mu = 0, sigma = 1, a = 2, b = -1, color = "green")

}
