% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Distribution curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mu is the mean}

\item{sigma}{sigma is the standard deviation}

\item{a}{a is the area}
}
\value{
Normal density curve
}
\description{
Normal Distribution curve
}
\examples{
myncurve(3, 7, 2)

}
