% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My Normal Curve Function}
\usage{
myncurve(mu, sigma, a = 0)
}
\arguments{
\item{mu}{the average of the normal}

\item{sigma}{the standard deviation of the normal}

\item{a}{upper bound of lower tail probability}
}
\value{
A plot of the density curve of the normal and a list contatining probability info
}
\description{
My Normal Curve Function
}
\examples{
myncurve(0, 1)
myncurve(5.5, 2.3, 6)
}
