% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My Normal Curve Generator}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean}

\item{sigma}{The standard deviation}

\item{a}{The a such that P(Y<=a)}
}
\value{
A plot of the curve with the area P(Y<=a) filled and the probability outputted to the command line.
}
\description{
My Normal Curve Generator
}
\examples{
myncurve(10,5,6)
}
