% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{sigma}{- the standard deviation of the distribution}

\item{a}{- the right end point of the area that is being summed}

\item{mu-}{the average of the distribution}
}
\value{
a graph with the area from -infinity to a shaded in.
}
\description{
This function takes the average, standard deviation and plots a normal graph
}
\details{
Using the parameter a, the area from -infinity to a is shaded and it's area is computed
}
\examples{
myncurve(5, 10, 6)
}
