% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot a curve, shade the given area and calculate the probability}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The Standard Deviation}

\item{sigma}{Given Value}

\item{a}{}
}
\value{
A shaded curve with its probability
}
\description{
The mean
}
\examples{
mu = 5, sigma = 2, a = 3, myncurve(mu, sigma, a)
}
