% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{Probability taht x = a, where P(X<=a)}

\item{mu}{mean of normal distribution}

\item{sigma}{standard deviation of normal distribution}
}
\value{
plot showing the normal curve w/prob of x <= a
}
\description{
myncurve
}
\examples{
myncurve(mu=10,sigma=2, a=6)

}
