% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{myncurve}
\alias{myncurve}
\title{Normal plot}
\usage{
myncurve(mu = 10, sigma = 5, a = 6)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{x value}
}
\value{
A normal distribution with shaded area
}
\description{
Normal plot
}
\examples{
\dontrun{myncurve(10,5,6)}
}
