% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mu}

\item{sigma}{sigma}

\item{a}{a}
}
\value{
Returns the curve
}
\description{
Title
}
\examples{
myncurve(2,2,2)
}
