% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot Normal Distribution Curve and Shade Area}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sigma}{Standard deviation of the normal distribution.}

\item{a}{The upper limit for shading under the curve.}
}
\value{
a list containing:
' \describe{
  \item{mu}{Mean of the distribution used.}
  \item{sigma}{Standard deviation of the distribution used.}
  \item{area}{The calculated probability (area under the curve) up to `a`.}
}
}
\description{
This function plots a normal distribution curve with a given mean (`mu`)
and standard deviation (`sigma`), then shades the area under the curve
up to a specified value `a`. It also computes the probability (area)
corresponding to P(X ≤ a).
}
\examples{
# Example: Plot N(0,1) and shade the area up to x = 1.5
myncurve(mu = 0, sigma = 1, a = 1.5)

}
