% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title Normal Distribution Plot}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{upper limit of lower tail probability}
}
\value{
plot shaded with lower tail probability
}
\description{
Title Normal Distribution Plot
}
\examples{
myncurve(1,1,1)
}
