% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{MynCurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{a}{}
}
\value{
A graph with the area shaded in green
}
\description{
MynCurve
}
\examples{
myncurve(4,2,5), where 4 = mean, 2 = sd, and 5 = a
}
