% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My Normal Distribution Function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{a vector representing the mean}

\item{sigma}{a vector representing the standard deviation}

\item{x}{a vector in which is value represents the upper limit for the x-axis}
}
\value{
a curve with the area shaded and plotting the value of the area
}
\description{
Creates a curve of a normal distribution function and plots and shaded the area
}
\details{
This is one of the functions in this package for the MATH4753 class. It takes the values of the vectors and plugs them into the equation. It creates a curve of a normal distribution, and shades the area created by the specific vectors and adds the value of the area to the curve.
}
\examples{
x=5
mu=3
sigma=6
myncurve(x,mu,sigma)
}
