% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean of the normal distribution}

\item{sigma}{standard deviation of the normal distribution}

\item{a}{X coordination you want to find P(Y<=a)}
}
\value{
a curve of the normal distribution and the area whose probability you are finding
}
\description{
myncurve
}
\examples{
myncurve(0,1,0)
}
