% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My normal distribution plotter}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{a number, the mean of the normal distribution}

\item{sigma}{a number, the standard deviation of the normal distribution}

\item{a}{a number, the point up to which to calculate the lower tail}
}
\value{
a plot with the lower tail shaded and a list containing the probability for a value to be in that area
}
\description{
Plots a normal distribution with the given parameters
}
\details{
Given the mean, standard deviation and a value to evaluate against, plots the curve and shades the lower tail, then returns the probability for a value to be in the lower tail.
}
\examples{
myncurve(mu=10,sigma=5,a=6)
}
