% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNCurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{your mean}

\item{sigma}{your standard deviation}

\item{a}{where you wanna stop your prob. estimate; P(-7<Y<a)}
}
\value{
a graph that shows the probability
}
\description{
myncurve
}
\examples{
myNCurve(10,5,9); returns a graph
}
