% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean}

\item{sigma}{The standarad deviation}

\item{a}{The point used for the probability of a certain set of values.}
}
\value{
A Normal curve with the probability region shaded and also returns
the value of the probability
}
\description{
myncurve
}
\examples{
myncurve(5, 5, 5)
}
