% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot a Normal Distribution Curve and Return Parameters}
\usage{
myncurve(mu, sigma)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sigma}{Standard deviation of the normal distribution.}
}
\value{
A list containing the mean (`mu`), standard deviation (`sigma`), and calculated probability.
}
\description{
Plot a Normal Distribution Curve and Return Parameters
}
\examples{
\dontrun{myncurve(mu = 0, sigma = 1)}
}
