% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Curve Visual and Probability Generator}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mu of the normal distribution}

\item{sigma}{sigma of the normal distribution}

\item{a}{a value for the probability calculation}
}
\value{
a list of mu, sigma, and probability
}
\description{
Curve Visual and Probability Generator
}
\examples{
\dontrun{myncurve(2, 3, 5)}
}
