% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Curve with Area Calculation}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean of the normal distribution}

\item{sigma}{The standard deviation of the normal distribution}

\item{a}{The value at which to calculate the probability P(X <= a)}
}
\value{
A list containing the mean, standard deviation, and calculated probability
}
\description{
Normal Curve with Area Calculation
}
\examples{
myncurve(0, 1, 1.5)
}
