% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(x, mu, sigma)
}
\arguments{
\item{x}{a vector}

\item{mu}{mean}

\item{sigma}{standard deviation}
}
\value{
a graph
}
\description{
myncurve
}
\examples{
myncurve(x,mu=2, sigma=4)
}
