% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu = 0, sigma = 1, a = 2)
}
\arguments{
\item{mu}{mean}

\item{sigma}{sd}

\item{a}{x = a}
}
\value{
Show curve and area from -inf to x = a
}
\description{
Title
}
\examples{
myncurve(mu = 0, sigma = 1, a = 2)
}
