% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyDistribution.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma)
}
\arguments{
\item{mu}{The mean of the normal probability distribution curve. Essentially defines the center of the curve of the distribution of the probability.}

\item{sigma}{The standard deviation of the normal probability distribution curve.}
}
\value{
a curve of the normal distribution (defined mean and standard deviation)
}
\description{
Returns a curve for a distribution with parameters mu and sigma representing the mean and the standard deviation.
}
\examples{
myncurve(mu = 2, sigma = 1)
}
