% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNorm.R
\name{myNorm}
\alias{myNorm}
\title{Normal Experiment}
\usage{
myNorm(x, mu, sig, ...)
}
\arguments{
\item{x}{the mean}

\item{mu}{the standard deviation}

\item{sig}{more parameters can be read in}

\item{...}{}
}
\value{
a contour graph,
and the vector, the coordinates and mle is asked for
}
\description{
the sample vector
}
\examples{
myNorm(x = c(10,12,13,15,12,11,10), mu = seq(8, 16, length = 1000), sig = seq(0.1, 4, length = 1000), lwd = 2, labcex = 1)
}
