% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNRML.R
\name{myNRML}
\alias{myNRML}
\title{Newton Rhapson maximum likelihood}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{the initial x point at which to start analysing the graph to see where it intersects x-axis}

\item{delta}{the difference between x points analysed}

\item{llik}{function you want to implement. Must be a log base e function}

\item{xrange}{the range of the x axis for the log likelihood graph}

\item{parameter}{the parameter that is to be estimate eg. mu, sigma, lambda}
}
\value{
returns two graphs, first the log likelihood graph of the parameter of interest. Second the derivative of the log likelihood graph that shows the different x points analysed and the final x point at which the graph intersects the x axis
}
\description{
Newton Rhapson maximum likelihood
}
\examples{
  \dontrun{myNRML(x0=1,delta=0.000001,llik=function(x) log(dpois(4,x)*dpois(5,x)),xrange=c(0,20),parameter="lambda")}
}
