% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab10.R
\name{myNRML}
\alias{myNRML}
\title{Newton-Raphson maximum likelihood approximation}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{parameter}{}
}
\value{

}
\description{
approximate a parameter of interest
This needs initial x guess
}
\examples{
myNRML(x0=1,delta=0.000001,llik=function(x) log(dpois(12,x)*dpois(10,x)),xrange=c(0,20),parameter="lambda" )
}
