% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab10.R
\name{myNRML}
\alias{myNRML}
\title{Newton-Raphson maximum likelihood approximation}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{initial guess of root}

\item{delta}{acceptable difference for solution convergence}

\item{llik}{likelihood function whose root to be estimated}

\item{xrange}{set range for x in the graphic}

\item{parameter}{set parameter as display on graphic}
}
\value{
root estimated by Newton Raphson for maximum likelihood estimation
}
\description{
approximate a parameter of interest
This needs initial x guess
}
\examples{
myNRML(x0=1,delta=0.000001,llik=function(x) log(dpois(12,x)*dpois(10,x)),xrange=c(0,20),parameter="lambda" )
}
