% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNRML.R
\name{myNRML}
\alias{myNRML}
\title{Newton-Raphson Maximum Likelihood Estimation}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{Initial guess for the parameter of interest.}

\item{delta}{A small increment in the parameter to estimate its derivative.}

\item{llik}{A function that computes the log-likelihood of the observed data given the parameter.}

\item{xrange}{The range of x values over which the log-likelihood and its derivative are plotted.}

\item{parameter}{The name of the parameter being estimated, used for labeling plots.}
}
\value{
Returns a list containing two elements: `x` which is a vector of the
        parameter estimates at each iteration, and `y` which is a vector of the
        derivative values of the log-likelihood function at the corresponding estimates.
        The function also produces two side-by-side plots: one for the log-likelihood
        function and one for its derivative, with the iterations of the algorithm overlaid.
}
\description{
Performs the Newton-Raphson method to find the maximum likelihood estimator
for a parameter given a log-likelihood function. It iteratively finds better
approximations to the roots (i.e., the maximum likelihood estimates) of a real-valued function.
}
\examples{
myNRML(x0 = 1, delta = 0.000001,
       llik = function(x) log(dpois(4, x) * dpois(6, x) * dpois(7, x) *
                               dpois(6, x) * dpois(5, x)),
       xrange = c(0, 30), parameter = "lambda")

}
