% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyNrml.R
\name{myNRML}
\alias{myNRML}
\title{MyNRML}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{parameter}{}
}
\value{

}
\description{
MyNRML
}
\examples{
myNRML(x0=c(4,6,7,6,5),delta=0.001,llik=function(x) log(dpois(12,x)*dpois(10,x)),xrange=c(0,20),parameter="lambda" )
}
