% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myquadplot.R
\name{myplot}
\alias{myplot}
\title{Quadratic function}
\usage{
myplot(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Intercept of the quadratic function
}
\description{
Quadratic function
}
\examples{
myplot(x = 1:10)

}
