% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlot.R
\name{myPlot}
\alias{myPlot}
\title{Helper function for plotting}
\usage{
myPlot(x, coef, xk)
}
\arguments{
\item{x}{Numeric value of x}

\item{coef}{Numeric vector of coefficient values}

\item{xk}{Numeric value of $x_k$}
}
\description{
Makes plotting piece-wise regressions easier, created for lab 4
}
\examples{
coef = c(1, 2 ,3)
myPlot(0, coef = coef, xk = 18)

}
