% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlot.R
\name{myPlot}
\alias{myPlot}
\title{Helper function for plotting}
\usage{
myPlot(x, coef, xk)
}
\arguments{
\item{x}{Numeric value}

\item{coef}{Numeric vector}

\item{xk}{Numeric value}
}
\description{
Makes plotting piecewise regressions easier
}
\examples{
coef = c(1, 2 ,3)
myPlot(0, coef = coef, xk = 18)

}
