% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{Computes the p-value for a given t-statistic and plots the t-distribution}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{Observed t-statistic.}

\item{xmax}{Maximum x-axis value for the plot.}

\item{n}{Sample size.}

\item{alpha}{Significance level (default is 0.05).}
}
\value{
A list containing:
\item{q}{Critical t-values for the given alpha.}
\item{pvalue}{The computed p-value for the observed t-statistic.}
}
\description{
Computes the p-value for a given t-statistic and plots the t-distribution
}
