% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{P value function}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{tcalc}

\item{xmax}{xmax on the plot}

\item{n}{sample size}

\item{alpha}{alpha}
}
\description{
P value function
}
\examples{
mypvalue(3.456)
}
