% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{My P Value}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{alpha}{}
}
\value{
P value
}
\description{
This helps us determine what the P value is.
}
\examples{

mypvalue(tcalc, xmax=4, n=30, alpha=0.05)

}
