% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{p-value Function}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{a vector of t-calc}

\item{xmax}{vector of the maximum value of x}

\item{n}{the sample size}

\item{alpha}{1 - confidence level}
}
\value{
a vector with the value of p-value, a vector of the value of the start of the rejection regions along with a curve that shows the two rejection regions and the acceptance region
}
\description{
Finds the two rejection regions and the acceptance region along with the p-value
}
\details{
This function calculates alpha/2, then plots the t distance. This function then sets up points on the polygon to the right and left, then shades in the polygon defined by the line segments. It makes quantiles then plots the cut off t value and adds some annotation to the graph.
}
\examples{
set.seed(55)
x1=rnorm(30,mean=25,sd=5)
bootpval(x=x1,mu0=24,test="two")
}
