% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{course 4753}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{x}{to find the p value}
}
\value{
value of p value
}
\description{
to find the p value.
}
\examples{
set.seed(55);x1=rnorm(30,mean=25,sd=5)
tcalc=(mean(x1)-23)/(sd(x1)/sqrt(30))
mypvalue(tcalc,n=30,alpha=0.05)



}
