% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{File Reading Function}
\usage{
myread(csv)
}
\arguments{
\item{csv}{}
}
\value{
the data read in from the csv file
}
\description{
Takes a CSV file, read in the data
}
\examples{
csv = "spruce.csv"; myread(csv)
}
