% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{CSV file reader}
\usage{
myread(csv)
}
\arguments{
\item{csv}{}
}
\value{
data frame containing data
}
\description{
This is to take any csv file in a directory and read it into R.
}
\examples{
dird="E:\\\\OneDrive\\\\MATH4753\\\\DATAxls\\\\"; myread("NAME.csv")
}
