% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRead.R
\name{myread}
\alias{myread}
\title{CSV Reader (Lab 2 Version)}
\usage{
myread(csv)
}
\arguments{
\item{csv}{The name of the CSV file being read}
}
\value{
data frame that contains the data
}
\description{
CSV Reader (Lab 2 Version)
}
\examples{
mpg.df=myread("EPAGAS.csv")
}
