% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{CSV File Reader}
\usage{
myread(csv, dird)
}
\arguments{
\item{csv}{csv file name (including.csv)}

\item{dird}{directory where the csv file is located}
}
\value{
a data frame of the csv file
}
\description{
CSV File Reader
}
\examples{
\dontrun{spruce.df=myread("SPRUCE.csv", 
        "/Users/tiffanynguyen/Documents/College/Sophomore/Spring_2025/Applied_Statistical_Methods/MATH4753nguy0850/labs/lab4/DATAxls/")}
}
