% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{Title}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{size}

\item{iter}{number of iterations}

\item{time}{times run}
}
\value{
Barplot of n run iter times
}
\description{
Title
}
\examples{
mysample(10,1,1)
}
