% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySimulation.R
\name{mySimulation}
\alias{mySimulation}
\title{Binomial simulation plotting function}
\usage{
mySimulation(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{number of size}

\item{n}{probability}

\item{p}{}
}
\value{
A table along with a barplot of the simulation
}
\description{
number of time for iteration
}
\examples{
iter = 10,000, n = 5, p = 0.5 mybin(iter, n, p)
}
