% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Optimal Number of Tickets to Sell}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats available on the flight}

\item{gamma}{Acceptable risk of overbooking}

\item{p}{Probability that a given passenger shows up}
}
\value{
Returns requirements listed in project 1.
}
\description{
This function calculates the optimal number of tickets to sell for a flight
}
\examples{
nticketresult <- ntickets(400, 0.02, 0.95)
print(nticketresult)

}
