% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{is the number of flight seats}

\item{gamma}{probability plane will be overbooked}

\item{p}{probability passenger will show}
}
\value{
two plots of continuous and discrete gamma functions
}
\description{
ntickets function
}
\examples{
ntickets(N = 200, gamma = 0.02, p = 0.95)

}
