% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Airline ticket overbooking simulator}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95, layout = FALSE)
}
\arguments{
\item{N}{the total seats on the flight}

\item{gamma}{probability that every ticket holder shows up to board}

\item{p}{probability that a person who is sold a ticket shows up to board}

\item{layout}{include both generated plots in a single window. use this if running outside Rmd}
}
\value{
two plots: one showing distribution values used to calculate the number of tickets to sell with a binomial distribution, and the other with a normal approximation. Also returns a list containing these calculated values, N, gamma, and p.
}
\description{
Uses binomial and normal variables to test the number of tickets an airline should sell given parameters
}
\examples{
ntickets()
ntickets(N=400, gamma=0.02, p=0.95)
}
