% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats on the plane}

\item{gamma}{Probability of overbooking}

\item{p}{Probability that a passenger "shows"}
}
\value{
Two graphs of the objective functions vs n (discrete and continuous) and a named list
}
\description{
ntickets
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.95)
}
