% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats available}

\item{gamma}{probability of overbooking}

\item{p}{probability that a passenger will show}
}
\value{
A list of nd, nc, N, p, and gamma; and two graphs, a discrete and continuous curve of optimal tickets to sell
}
\description{
ntickets
}
\examples{
ntickets(400,.02, .95)

}
