% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Overbooking}
\usage{
ntickets(N = 400, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{The number of seats on the plane}

\item{gamma}{The probability of overbooking}

\item{p}{The probability of someone showing up}
}
\value{
The number of tickets that should be sold
for both the discrete and continuous case given p
and gamma. Returns a discrete and continuous,
normal approximated plot of objective function vs
the number of tickets sold, where the objective
function describes how overbook the plane is. If
the objective function is equal to 0, then the
number of tickets at that value is the number of
tickets that minimizes overbooking.
}
\description{
Overbooking
}
\examples{
\dontrun{ntickets(200,0.02,0.95)}

}
