% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Number of Tickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{The number of seats on a plane}

\item{gamma}{The probability of overbooking}

\item{p}{The probability that someone will show up for their flight}
}
\value{
2 plots and a named list of nd, nc, N, and p
}
\description{
Number of Tickets
}
\examples{
\dontrun{ntickets(N=400,gamma = 0.02, p = 0.95)}
}
