% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Calculate number of tickets to sell for a flight given overbooking risk}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Integer: Number of seats on the flight.}

\item{gamma}{Numeric: Probability that the flight is truly overbooked.}

\item{p}{Numeric: Probability that a passenger shows up.}
}
\value{
A named list containing:
\describe{
  \item{nd}{Number of tickets using discrete distribution.}
  \item{nc}{Number of tickets using normal approximation.}
  \item{N}{Number of seats.}
  \item{p}{Show up probability.}
  \item{gamma}{Overbooking probability.}
}
}
\description{
This function computes the number of tickets to be sold for a flight with
N seats, given the probability `p` that a passenger shows up and
gamma, the probability that the flight will be overbooked.
}
\details{
The function calculates two ticket sale limits:
\itemize{
  \item \code{nd}: based on the discrete (binomial) distribution
  \item \code{nc}: based on the normal approximation
}
It also plots the objective functions for both methods.
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.95)

}
