% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Determine Number of Tickets to be Sold for a Flight}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats available in the flight.}

\item{gamma}{Maximum allowable probability that the plane is overbooked.}

\item{p}{Probability that a ticket holder shows up for the flight.}
}
\value{
A named list containing:
\itemize{
  \item \code{nd}: Number of tickets to be sold as calculated using the discrete distribution.
  \item \code{nc}: Number of tickets to be sold as calculated using the normal approximation.
  \item \code{N}: Number of seats available in the flight.
  \item \code{gamma}: Maximum allowable probability that the plane is overbooked.
  \item \code{p}: Probability that a ticket holder shows up for the flight.
}

The function also generates two plots of the Objective function versus n. The Objective function
is derived from the defining equation set to zero (e.g., 1-gamma-pnorm(...) = 0). One plot is for
the discrete case and the other for the continuous (normal approximation) case.
}
\description{
The function `ntickets` calculates the number of tickets to sell for a flight with `N` seats,
ensuring that the probability the plane is overbooked does not exceed `gamma`. The calculation
is done in two ways:
1) Using the appropriate discrete distribution.
2) Using the normal approximation.
}
\examples{
data <- ntickets(N=200, gamma=0.02, p=0.95)
print(data)

}
