% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{nTickets}
\usage{
ntickets(N = 200, gamma = 0.02, prob = 0.95)
}
\arguments{
\item{N}{= Maximum capacity of the plane}

\item{gamma}{= The probability that the plane is truly overbooked}

\item{prob}{= The probability of someone not showing up to the plane}
}
\value{
A named list of the discrete and continuous optimal numbers for the overbooking problem, along with the parameters of the function.
}
\description{
nTickets
}
\examples{
ntickets(N=400,gamma = 0.02, p = 0.95)
}
