% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Determine Optimal Number of Tickets for Overbooking}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats on the flight.}

\item{gamma}{Allowable probability of overbooking.}

\item{p}{Probability that a passenger shows up.}
}
\value{
A named list with elements:
\describe{
  \item{nd}{Optimal tickets using the discrete binomial method.}
  \item{nc}{Optimal tickets using the normal approximation.}
  \item{N}{The input number of seats.}
  \item{p}{The input probability of show.}
  \item{gamma}{The input overbooking probability.}
}
}
\description{
Calculates the number of tickets to sell given the number of seats (N),
the probability that a passenger shows up (p), and the allowable overbooking probability (gamma).
The function computes the optimal number of tickets using:
A discrete binomial approach and,
A normal approximation.
It also produces two separate plots: one for the discrete objective function and one for the continuous objective.
}
\examples{
\dontrun{
  results <- ntickets(N = 400, gamma = 0.02, p = 0.95)
  print(results)
}

}
