% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets - a function that calculates optimal passengers booked for a flight}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats available}

\item{gamma}{Desired probability that a flight is overbooked}

\item{p}{Probability that a passenger will no-show}
}
\value{
Prints requested values, and also plots the results
}
\description{
ntickets - a function that calculates optimal passengers booked for a flight
}
\examples{
\dontrun{ntickets(200,.02,0.95)}
}
