% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{calculates the number of tickets to be sold}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats in the flight}

\item{gamma}{probability that the plane will truly be overbooked}

\item{p}{probability of a "show"}
}
\value{
the number of tickets to be sold
}
\description{
calculates the number of tickets to be sold
}
\examples{
\dontrun{functionCall = ntickets(N = 400, gamma = 0.02, p = 0.95)}
}
