% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{the total number of trials}

\item{gamma}{the acceptable amount of pain}

\item{p}{the probability of success}
}
\value{
a named list and plots
}
\description{
ntickets
}
\examples{
ntickets(200,0.02,0.95)
}
