% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj1func.R
\name{ntickets}
\alias{ntickets}
\title{Title Optimize Ticket Selling}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{capacity of plane}

\item{gamma}{probability plane will be overbooked}

\item{p}{probability plane}
}
\value{
a list of recommended tickets to sell using discrete and continuous functions
}
\description{
Title Optimize Ticket Selling
}
\examples{
\dontrun{ntickets(N=200,gamma = 0.02, p = 0.95)}
}
