% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{Scatterplot and Histogramp graph}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{variables grphed on the x axis}

\item{y}{the variables graphed on the y axis}

\item{xlab}{The name of the x variables. If this parameter is used the name is plotted on the x axis}

\item{ylab}{The name of the y variables. If this parameter is used the name is plotted on the y axis}
}
\value{
a histogram plotted on the outside of a scatter plot
}
\description{
A function that plots scatter plot and then graphs a histogram on the edges of the scatter plot
}
\details{
This function is used with the with function
}
\examples{
with(ddt, scatterhist(LENGTH,WEIGHT, xlab="LENGTH"))
}
