% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{Scatter histogram}
\usage{
scatterhist(x = 5, y = 5, xlab = "0", ylab = "1")
}
\arguments{
\item{x}{X variable as a vector}

\item{y}{Y variable as a vector}

\item{xlab}{x label}

\item{ylab}{y label}
}
\value{
A detailed scatter plot using the specified data
}
\description{
Scatter histogram
}
\examples{
\dontrun{with(ddt, scatterhist(LENGTH,WEIGHT, xlab="LENGTH"))}
}
